Make A SQL statement

Not meant to be called directly by the user.

stat_make(reg, study, min_abs_cor, max_num, targets, type = "mir")

Arguments

reg
A character vector of one or more regulator ID.

study
A character vector of The Cancer Genome Atlas (TCGA) study identifiers. To view the available studies in TCGA project, https://tcga-data.nci.nih.gov/docs/publications/tcga. When left to default NULL all available studies will be included.

min_abs_cor
A numeric, an absolute correlation minimum between 0 and 1 for each mir.

max_num
An integer, maximum number of features to show for each mir in each study.

targets
A character vector of gene symbol names.

type
A character string. Either 'mir' of 'tf'. Used to define columns and tables names.

Value

A character string

Examples

stat_make(reg = 'hsa-let-7g',
          study = 'STES')

#> [1] "SELECT cor_mir.mirna_base, cor_mir.feature, cor_mir.STES FROM cor_mir WHERE cor_mir.mirna_base=\"hsa-let-7g\""

          
stat_make(reg = 'hsa-let-7g',
          study = 'STES',
          min_abs_cor = .3)

#> [1] "SELECT cor_mir.mirna_base, cor_mir.feature, cor_mir.STES FROM cor_mir WHERE cor_mir.mirna_base=\"hsa-let-7g\" AND ABS(cor_mir.STES) > 30"

          
stat_make(reg = 'hsa-let-7g',
          study = 'STES',
          min_abs_cor = .3,
          max_num = 5)

#> [1] "SELECT cor_mir.mirna_base, cor_mir.feature, cor_mir.STES FROM cor_mir WHERE cor_mir.mirna_base=\"hsa-let-7g\" AND ABS(cor_mir.STES) > 30 ORDER BY ABS(cor_mir.STES) DESC LIMIT 5"