Get microRNA correlations from cRegulome.db

This function access the sqlite database file which is obtained by running get_db. Basically, the function provides ways to query the database to the correlation data of the microRNAs of interest. The function returns an error if the database file cRegulome.db is not in the working directory.

Usage

get_mir(conn, mir, study, min_abs_cor, max_num, targets_only = FALSE, targets)

Arguments

conn

A connection such as this returned by dbConnect

mir

A required character vector of the microRNAs of interest. These are the miRBase ID which are the official identifiers of the widely used miRBase database, http://www.mirbase.org/.

study

A character vector of The Cancer Genome Atlas (TCGA) study identifiers. To view the available studies in TCGA project, https://tcga-data.nci.nih.gov/docs/publications/tcga. When left to default NULL all available studies will be included.

min_abs_cor

A numeric, an absolute correlation minimum between 0 and 1 for each mir.

max_num

An integer, maximum number of features to show for each mir in each study.

targets_only

A logical whether restrict the output to the recognized target features.

targets

A character vector of gene symbol names.

Value

A tidy data.frame of four columns. mirna_base is the microRNA miRBase IDs, feature is the features/genes, cor

is the corresponding expression correlations and study is TCGA study ID.

Examples

# locate the testset file and connect
fl <- system.file('extdata', 'cRegulome.db', package = 'cRegulome')
conn <- RSQLite::dbConnect(RSQLite::SQLite(), fl)

# get microRNA correlations in all studies
get_mir(conn,
        mir = 'hsa-let-7g')
#>     mirna_base   feature   cor study
#> 1   hsa-let-7g      ABL2  0.11  STES
#> 2   hsa-let-7g    ACTR10 -0.16  STES
#> 3   hsa-let-7g    ACVR1C -0.13  STES
#> 4   hsa-let-7g   ADAMTS1  0.27  STES
#> 5   hsa-let-7g   ADAMTS5  0.20  STES
#> 6   hsa-let-7g     ADCY9  0.18  STES
#> 7   hsa-let-7g     ADRB1  0.11  STES
#> 8   hsa-let-7g     ADRB2 -0.12  STES
#> 9   hsa-let-7g      AFF2 -0.10  STES
#> 10  hsa-let-7g     AKAP5  0.10  STES
#> 11  hsa-let-7g     AKAP6  0.12  STES
#> 12  hsa-let-7g      AKT2 -0.14  STES
#> 13  hsa-let-7g   ANKRD12  0.11  STES
#> 14  hsa-let-7g   ANKRD28  0.16  STES
#> 15  hsa-let-7g   ANKRD46 -0.18  STES
#> 16  hsa-let-7g     AP4E1 -0.11  STES
#> 17  hsa-let-7g    APPBP2  0.13  STES
#> 18  hsa-let-7g  ARHGAP20  0.28  STES
#> 19  hsa-let-7g  ARHGEF15  0.39  STES
#> 20  hsa-let-7g  ARHGEF38  0.10  STES
#> 21  hsa-let-7g     ARL4D -0.15  STES
#> 22  hsa-let-7g   ARL6IP6 -0.18  STES
#> 23  hsa-let-7g    ATP2A2 -0.12  STES
#> 24  hsa-let-7g    ATP2B4  0.17  STES
#> 25  hsa-let-7g   ATXN7L2 -0.12  STES
#> 26  hsa-let-7g   ATXN7L3 -0.15  STES
#> 27  hsa-let-7g     AURKB -0.20  STES
#> 28  hsa-let-7g      AVEN -0.18  STES
#> 29  hsa-let-7g    B3GNT7  0.10  STES
#> 30  hsa-let-7g       BBX  0.13  STES
#> 31  hsa-let-7g    BCL2L1  0.12  STES
#> 32  hsa-let-7g     BCL7A -0.17  STES
#> 33  hsa-let-7g      BNC2  0.14  STES
#> 34  hsa-let-7g      BRF2 -0.12  STES
#> 35  hsa-let-7g    BTF3L4 -0.11  STES
#> 36  hsa-let-7g  C11orf84 -0.12  STES
#> 37  hsa-let-7g   C1GALT1  0.12  STES
#> 38  hsa-let-7g C20orf194  0.12  STES
#> 39  hsa-let-7g   C2orf88  0.14  STES
#> 40  hsa-let-7g   C3orf52  0.12  STES
#> 41  hsa-let-7g   C8orf58 -0.12  STES
#> 42  hsa-let-7g   CACNA1D  0.13  STES
#> 43  hsa-let-7g     CALM1  0.11  STES
#> 44  hsa-let-7g   CASKIN1 -0.12  STES
#> 45  hsa-let-7g      CBX2 -0.12  STES
#> 46  hsa-let-7g      CCL7 -0.10  STES
#> 47  hsa-let-7g      CCNF -0.17  STES
#> 48  hsa-let-7g     CCNJL -0.11  STES
#> 49  hsa-let-7g     CD276 -0.16  STES
#> 50  hsa-let-7g    CDC14B  0.11  STES
#> 51  hsa-let-7g    CDC25A -0.11  STES
#> 52  hsa-let-7g  CDC42SE1 -0.23  STES
#> 53  hsa-let-7g     CDCA8 -0.13  STES
#> 54  hsa-let-7g    CERCAM -0.10  STES
#> 55  hsa-let-7g      CHRD  0.15  STES
#> 56  hsa-let-7g    CLASP2  0.14  STES
#> 57  hsa-let-7g      CLP1 -0.16  STES
#> 58  hsa-let-7g    CNOT6L  0.13  STES
#> 59  hsa-let-7g      COIL -0.11  STES
#> 60  hsa-let-7g   COL11A1 -0.12  STES
#> 61  hsa-let-7g   COL14A1  0.15  STES
#> 62  hsa-let-7g   COL15A1  0.19  STES
#> 63  hsa-let-7g    COL4A1  0.13  STES
#> 64  hsa-let-7g    COL4A2  0.15  STES
#> 65  hsa-let-7g    COL4A6  0.14  STES
#> 66  hsa-let-7g       CPD  0.14  STES
#> 67  hsa-let-7g     CPEB4  0.10  STES
#> 68  hsa-let-7g      CRBN  0.20  STES
#> 69  hsa-let-7g     CRTAP  0.13  STES
#> 70  hsa-let-7g      CRY2  0.15  STES
#> 71  hsa-let-7g      CTSC -0.17  STES
#> 72  hsa-let-7g   CXorf36  0.36  STES
#> 73  hsa-let-7g     CYTH3  0.14  STES
#> 74  hsa-let-7g     DAGLA  0.14  STES
#> 75  hsa-let-7g     DAPK1  0.21  STES
#> 76  hsa-let-7g    DCAF15 -0.21  STES
#> 77  hsa-let-7g     DCAF8  0.10  STES
#> 78  hsa-let-7g       DDN -0.19  STES
#> 79  hsa-let-7g      DLC1  0.26  STES
#> 80  hsa-let-7g       DMD  0.19  STES
#> 81  hsa-let-7g      DMP1 -0.10  STES
#> 82  hsa-let-7g      DNA2 -0.13  STES
#> 83  hsa-let-7g    DNAJA2 -0.11  STES
#> 84  hsa-let-7g    DNAJC1  0.17  STES
#> 85  hsa-let-7g     DNAL1  0.10  STES
#> 86  hsa-let-7g     DOCK3  0.24  STES
#> 87  hsa-let-7g      DPH3  0.15  STES
#> 88  hsa-let-7g    DPYSL3  0.18  STES
#> 89  hsa-let-7g      DSG3 -0.19  STES
#> 90  hsa-let-7g     DUSP7 -0.19  STES
#> 91  hsa-let-7g     DUSP9 -0.15  STES
#> 92  hsa-let-7g      DVL3 -0.15  STES
#> 93  hsa-let-7g    DYRK1A  0.11  STES
#> 94  hsa-let-7g     DYRK2  0.11  STES
#> 95  hsa-let-7g      E2F2 -0.13  STES
#> 96  hsa-let-7g       EDA  0.10  STES
#> 97  hsa-let-7g     EDEM3  0.11  STES
#> 98  hsa-let-7g      EDN1  0.24  STES
#> 99  hsa-let-7g      EEA1 -0.10  STES
#> 100 hsa-let-7g     EGLN2  0.11  STES
#> 101 hsa-let-7g    EIF4G2 -0.10  STES
#> 102 hsa-let-7g    ELOVL4 -0.12  STES
#> 103 hsa-let-7g     EPHA3  0.13  STES
#> 104 hsa-let-7g     ERCC4 -0.10  STES
#> 105 hsa-let-7g     ERCC6 -0.10  STES
#> 106 hsa-let-7g    ERGIC1  0.11  STES
#> 107 hsa-let-7g     ESPL1 -0.16  STES
#> 108 hsa-let-7g     ETNK2 -0.12  STES
#> 109 hsa-let-7g      EZH1  0.20  STES
#> 110 hsa-let-7g      EZH2 -0.13  STES
#> 111 hsa-let-7g  FAM103A1 -0.16  STES
#> 112 hsa-let-7g   FAM104A -0.12  STES
#> 113 hsa-let-7g   FAM118A  0.15  STES
#> 114 hsa-let-7g       FAS -0.10  STES
#> 115 hsa-let-7g    FBXL19 -0.20  STES
#> 116 hsa-let-7g    FBXO32  0.12  STES
#> 117 hsa-let-7g     FGF11 -0.20  STES
#> 118 hsa-let-7g    FNDC3A  0.14  STES
#> 119 hsa-let-7g     FOXN3  0.22  STES
#> 120 hsa-let-7g     FOXP1  0.24  STES
#> 121 hsa-let-7g     FOXP2  0.14  STES
#> 122 hsa-let-7g      FRS2  0.15  STES
#> 123 hsa-let-7g      FZD3 -0.10  STES
#> 124 hsa-let-7g      FZD4  0.19  STES
#> 125 hsa-let-7g      GAB2  0.15  STES
#> 126 hsa-let-7g      GALC  0.12  STES
#> 127 hsa-let-7g      GALE  0.10  STES
#> 128 hsa-let-7g    GALNT1 -0.15  STES
#> 129 hsa-let-7g    GALNT4  0.17  STES
#> 130 hsa-let-7g      GAS7  0.21  STES
#> 131 hsa-let-7g     GDPD1  0.17  STES
#> 132 hsa-let-7g    GEMIN7 -0.10  STES
#> 133 hsa-let-7g     GIPC1 -0.14  STES
#> 134 hsa-let-7g      GNAL  0.11  STES
#> 135 hsa-let-7g       GNS  0.12  STES
#> 136 hsa-let-7g    GOLGA7 -0.10  STES
#> 137 hsa-let-7g    GOLT1B -0.12  STES
#> 138 hsa-let-7g   GRID2IP  0.14  STES
#> 139 hsa-let-7g     GTF2I -0.13  STES
#> 140 hsa-let-7g     HABP4  0.12  STES
#> 141 hsa-let-7g     HCFC2  0.11  STES
#> 142 hsa-let-7g       HDX  0.12  STES
#> 143 hsa-let-7g      HIC1  0.11  STES
#> 144 hsa-let-7g       HK2 -0.16  STES
#> 145 hsa-let-7g       HLF  0.21  STES
#> 146 hsa-let-7g     HMGA1 -0.12  STES
#> 147 hsa-let-7g      HTR4  0.14  STES
#> 148 hsa-let-7g      IGF1  0.12  STES
#> 149 hsa-let-7g     IGF1R -0.18  STES
#> 150 hsa-let-7g    IKBKAP -0.10  STES
#> 151 hsa-let-7g      IL13  0.34  STES
#> 152 hsa-let-7g      IL6R  0.12  STES
#> 153 hsa-let-7g     IMPG2  0.21  STES
#> 154 hsa-let-7g    INPP5A  0.16  STES
#> 155 hsa-let-7g      INSR  0.24  STES
#> 156 hsa-let-7g     INTS2 -0.11  STES
#> 157 hsa-let-7g     ITGB3  0.24  STES
#> 158 hsa-let-7g     KCMF1 -0.20  STES
#> 159 hsa-let-7g     KCNC4  0.13  STES
#> 160 hsa-let-7g     KCNQ4  0.20  STES
#> 161 hsa-let-7g    KCTD10  0.21  STES
#> 162 hsa-let-7g     KIFC2 -0.14  STES
#> 163 hsa-let-7g      KLF9  0.27  STES
#> 164 hsa-let-7g    KLHL23  0.24  STES
#> 165 hsa-let-7g     KLHL6  0.10  STES
#> 166 hsa-let-7g   KREMEN1 -0.14  STES
#> 167 hsa-let-7g     LAMP2 -0.15  STES
#> 168 hsa-let-7g      LCOR  0.16  STES
#> 169 hsa-let-7g      LDB3  0.15  STES
#> 170 hsa-let-7g     LIMD1  0.13  STES
#> 171 hsa-let-7g     LIMK2 -0.19  STES
#> 172 hsa-let-7g      LIPH  0.13  STES
#> 173 hsa-let-7g      LMLN -0.14  STES
#> 174 hsa-let-7g     LOXL3  0.10  STES
#> 175 hsa-let-7g     LRIG1  0.12  STES
#> 176 hsa-let-7g   LRRC10B  0.13  STES
#> 177 hsa-let-7g     LSM11  0.38  STES
#> 178 hsa-let-7g    LUC7L3  0.15  STES
#> 179 hsa-let-7g    MAP3K3  0.23  STES
#> 180 hsa-let-7g    MAP4K3  0.15  STES
#> 181 hsa-let-7g    MAP4K4 -0.10  STES
#> 182 hsa-let-7g     MAPK6 -0.19  STES
#> 183 hsa-let-7g      MBD2 -0.26  STES
#> 184 hsa-let-7g      MDFI -0.11  STES
#> 185 hsa-let-7g      MED6 -0.15  STES
#> 186 hsa-let-7g     MEF2C  0.17  STES
#> 187 hsa-let-7g     MEF2D  0.20  STES
#> 188 hsa-let-7g     MEX3A -0.10  STES
#> 189 hsa-let-7g     MGAT3  0.10  STES
#> 190 hsa-let-7g    MGAT4A  0.13  STES
#> 191 hsa-let-7g      MGLL  0.15  STES
#> 192 hsa-let-7g     MLXIP  0.11  STES
#> 193 hsa-let-7g      MTDH -0.10  STES
#> 194 hsa-let-7g     MTUS1  0.12  STES
#> 195 hsa-let-7g     NCOA1  0.16  STES
#> 196 hsa-let-7g     NCOR1  0.30  STES
#> 197 hsa-let-7g      NEFM -0.11  STES
#> 198 hsa-let-7g      NEK3  0.10  STES
#> 199 hsa-let-7g    NHLRC2  0.12  STES
#> 200 hsa-let-7g    NHLRC3  0.17  STES
#> 201 hsa-let-7g    NIPAL4 -0.13  STES
#> 202 hsa-let-7g   NKIRAS2 -0.13  STES
#> 203 hsa-let-7g       NLN -0.13  STES
#> 204 hsa-let-7g     NOVA1  0.18  STES
#> 205 hsa-let-7g      NRAS -0.18  STES
#> 206 hsa-let-7g       NRK  0.10  STES
#> 207 hsa-let-7g      NTN1  0.26  STES
#> 208 hsa-let-7g     OLFM4  0.11  STES
#> 209 hsa-let-7g   ONECUT2  0.10  STES
#> 210 hsa-let-7g     OTUD3  0.10  STES
#> 211 hsa-let-7g     P2RX1  0.11  STES
#> 212 hsa-let-7g     PANX2 -0.11  STES
#> 213 hsa-let-7g    PARD6B  0.16  STES
#> 214 hsa-let-7g      PBX1  0.14  STES
#> 215 hsa-let-7g    PCYT1B -0.11  STES
#> 216 hsa-let-7g     PDE12  0.10  STES
#> 217 hsa-let-7g     PDGFB  0.21  STES
#> 218 hsa-let-7g      PDPR  0.10  STES
#> 219 hsa-let-7g   PHACTR2  0.17  STES
#> 220 hsa-let-7g      PHC3  0.14  STES
#> 221 hsa-let-7g      PIGA -0.10  STES
#> 222 hsa-let-7g    PLA2G3 -0.10  STES
#> 223 hsa-let-7g   PLEKHO1  0.10  STES
#> 224 hsa-let-7g    PLXNA4  0.12  STES
#> 225 hsa-let-7g    PLXNC1  0.14  STES
#> 226 hsa-let-7g    PLXND1  0.21  STES
#> 227 hsa-let-7g    PMAIP1 -0.24  STES
#> 228 hsa-let-7g      PNKD  0.12  STES
#> 229 hsa-let-7g    POLR2D -0.15  STES
#> 230 hsa-let-7g    POU2F1  0.12  STES
#> 231 hsa-let-7g  PPARGC1A  0.12  STES
#> 232 hsa-let-7g    PRKAA2  0.11  STES
#> 233 hsa-let-7g  PSORS1C2 -0.10  STES
#> 234 hsa-let-7g    PTP4A2  0.12  STES
#> 235 hsa-let-7g     PTPRU  0.10  STES
#> 236 hsa-let-7g     PYGO2 -0.11  STES
#> 237 hsa-let-7g     RAB8B  0.12  STES
#> 238 hsa-let-7g      RAG1 -0.14  STES
#> 239 hsa-let-7g   RALGPS1  0.13  STES
#> 240 hsa-let-7g    RANBP2  0.16  STES
#> 241 hsa-let-7g     RBMS2  0.21  STES
#> 242 hsa-let-7g    RIMKLA  0.13  STES
#> 243 hsa-let-7g    RNF165 -0.14  STES
#> 244 hsa-let-7g     RNF44 -0.12  STES
#> 245 hsa-let-7g      RNF7 -0.15  STES
#> 246 hsa-let-7g   RPS6KA3  0.12  STES
#> 247 hsa-let-7g      RRM2 -0.21  STES
#> 248 hsa-let-7g     RSPO2  0.11  STES
#> 249 hsa-let-7g      RTKN -0.15  STES
#> 250 hsa-let-7g   RUNX1T1  0.18  STES
#> 251 hsa-let-7g    S100A8 -0.10  STES
#> 252 hsa-let-7g    SAP30L  0.11  STES
#> 253 hsa-let-7g     SCN4B  0.11  STES
#> 254 hsa-let-7g   SEC14L1  0.18  STES
#> 255 hsa-let-7g    SEC31B  0.20  STES
#> 256 hsa-let-7g SECISBP2L  0.12  STES
#> 257 hsa-let-7g    SEMA4F -0.15  STES
#> 258 hsa-let-7g     SENP2 -0.13  STES
#> 259 hsa-let-7g     SENP5 -0.13  STES
#> 260 hsa-let-7g    SESTD1  0.24  STES
#> 261 hsa-let-7g    SFMBT1  0.15  STES
#> 262 hsa-let-7g      SGCD  0.14  STES
#> 263 hsa-let-7g     SH2B3  0.14  STES
#> 264 hsa-let-7g      SIM2  0.10  STES
#> 265 hsa-let-7g   SLC16A9 -0.15  STES
#> 266 hsa-let-7g  SLC22A23 -0.13  STES
#> 267 hsa-let-7g   SLC30A6 -0.10  STES
#> 268 hsa-let-7g   SLC30A7  0.10  STES
#> 269 hsa-let-7g   SLC31A2 -0.15  STES
#> 270 hsa-let-7g   SLC38A9 -0.12  STES
#> 271 hsa-let-7g   SLC45A4  0.11  STES
#> 272 hsa-let-7g   SLC46A3  0.20  STES
#> 273 hsa-let-7g   SLC6A15 -0.18  STES
#> 274 hsa-let-7g     SNX16 -0.13  STES
#> 275 hsa-let-7g   SPATA13  0.12  STES
#> 276 hsa-let-7g      SPEG  0.15  STES
#> 277 hsa-let-7g    SPRYD4 -0.12  STES
#> 278 hsa-let-7g    SPTBN4  0.15  STES
#> 279 hsa-let-7g   STARD13  0.38  STES
#> 280 hsa-let-7g     STK40  0.15  STES
#> 281 hsa-let-7g      STRN -0.13  STES
#> 282 hsa-let-7g   SYNJ2BP  0.16  STES
#> 283 hsa-let-7g     SYT11  0.19  STES
#> 284 hsa-let-7g    TARBP2 -0.14  STES
#> 285 hsa-let-7g      TBX5  0.16  STES
#> 286 hsa-let-7g      TET3 -0.14  STES
#> 287 hsa-let-7g    TEX261 -0.12  STES
#> 288 hsa-let-7g    TGFBR3  0.21  STES
#> 289 hsa-let-7g     THBS1  0.21  STES
#> 290 hsa-let-7g     THOC2  0.12  STES
#> 291 hsa-let-7g     TMEM2  0.12  STES
#> 292 hsa-let-7g   TMEM41A -0.12  STES
#> 293 hsa-let-7g    TMEM65 -0.13  STES
#> 294 hsa-let-7g     TMOD2  0.24  STES
#> 295 hsa-let-7g     TMTC3 -0.22  STES
#> 296 hsa-let-7g  TNFRSF1B  0.13  STES
#> 297 hsa-let-7g    TNFSF9 -0.11  STES
#> 298 hsa-let-7g      TP53 -0.10  STES
#> 299 hsa-let-7g      TPP1  0.11  STES
#> 300 hsa-let-7g   TRAPPC1 -0.19  STES
#> 301 hsa-let-7g   TSC22D2 -0.10  STES
#> 302 hsa-let-7g   TSPAN18  0.14  STES
#> 303 hsa-let-7g    TTC39C -0.10  STES
#> 304 hsa-let-7g       TTL -0.10  STES
#> 305 hsa-let-7g     UCHL5 -0.15  STES
#> 306 hsa-let-7g     UHRF1 -0.21  STES
#> 307 hsa-let-7g     USP12  0.28  STES
#> 308 hsa-let-7g     USP20  0.15  STES
#> 309 hsa-let-7g     USP21 -0.14  STES
#> 310 hsa-let-7g     USP38  0.10  STES
#> 311 hsa-let-7g     USP44  0.11  STES
#> 312 hsa-let-7g     USP47  0.13  STES
#> 313 hsa-let-7g      UTRN  0.16  STES
#> 314 hsa-let-7g    VANGL2 -0.16  STES
#> 315 hsa-let-7g      VAV3 -0.18  STES
#> 316 hsa-let-7g     VGLL3  0.13  STES
#> 317 hsa-let-7g     VSNL1 -0.16  STES
#> 318 hsa-let-7g     WIPI2 -0.11  STES
#> 319 hsa-let-7g     XYLT1  0.10  STES
#> 320 hsa-let-7g     YPEL2  0.28  STES
#> 321 hsa-let-7g     ZBTB5 -0.10  STES
#> 322 hsa-let-7g     ZC3H3 -0.16  STES
#> 323 hsa-let-7g    ZNF197  0.19  STES
#> 324 hsa-let-7g    ZNF200 -0.11  STES
#> 325 hsa-let-7g    ZNF282 -0.18  STES
#> 326 hsa-let-7g   ZNF512B -0.12  STES
#> 327 hsa-let-7g    ZNF644  0.10  STES
#> 328 hsa-let-7g    ZNF879 -0.11  STES
#> 329 hsa-let-7g    ZSWIM5  0.11  STES

# get correlations in a particular study
get_mir(conn,
        mir = 'hsa-let-7g',
        study = 'STES')
#>     mirna_base   feature   cor study
#> 1   hsa-let-7g      ABL2  0.11  STES
#> 2   hsa-let-7g    ACTR10 -0.16  STES
#> 3   hsa-let-7g    ACVR1C -0.13  STES
#> 4   hsa-let-7g   ADAMTS1  0.27  STES
#> 5   hsa-let-7g   ADAMTS5  0.20  STES
#> 6   hsa-let-7g     ADCY9  0.18  STES
#> 7   hsa-let-7g     ADRB1  0.11  STES
#> 8   hsa-let-7g     ADRB2 -0.12  STES
#> 9   hsa-let-7g      AFF2 -0.10  STES
#> 10  hsa-let-7g     AKAP5  0.10  STES
#> 11  hsa-let-7g     AKAP6  0.12  STES
#> 12  hsa-let-7g      AKT2 -0.14  STES
#> 13  hsa-let-7g   ANKRD12  0.11  STES
#> 14  hsa-let-7g   ANKRD28  0.16  STES
#> 15  hsa-let-7g   ANKRD46 -0.18  STES
#> 16  hsa-let-7g     AP4E1 -0.11  STES
#> 17  hsa-let-7g    APPBP2  0.13  STES
#> 18  hsa-let-7g  ARHGAP20  0.28  STES
#> 19  hsa-let-7g  ARHGEF15  0.39  STES
#> 20  hsa-let-7g  ARHGEF38  0.10  STES
#> 21  hsa-let-7g     ARL4D -0.15  STES
#> 22  hsa-let-7g   ARL6IP6 -0.18  STES
#> 23  hsa-let-7g    ATP2A2 -0.12  STES
#> 24  hsa-let-7g    ATP2B4  0.17  STES
#> 25  hsa-let-7g   ATXN7L2 -0.12  STES
#> 26  hsa-let-7g   ATXN7L3 -0.15  STES
#> 27  hsa-let-7g     AURKB -0.20  STES
#> 28  hsa-let-7g      AVEN -0.18  STES
#> 29  hsa-let-7g    B3GNT7  0.10  STES
#> 30  hsa-let-7g       BBX  0.13  STES
#> 31  hsa-let-7g    BCL2L1  0.12  STES
#> 32  hsa-let-7g     BCL7A -0.17  STES
#> 33  hsa-let-7g      BNC2  0.14  STES
#> 34  hsa-let-7g      BRF2 -0.12  STES
#> 35  hsa-let-7g    BTF3L4 -0.11  STES
#> 36  hsa-let-7g  C11orf84 -0.12  STES
#> 37  hsa-let-7g   C1GALT1  0.12  STES
#> 38  hsa-let-7g C20orf194  0.12  STES
#> 39  hsa-let-7g   C2orf88  0.14  STES
#> 40  hsa-let-7g   C3orf52  0.12  STES
#> 41  hsa-let-7g   C8orf58 -0.12  STES
#> 42  hsa-let-7g   CACNA1D  0.13  STES
#> 43  hsa-let-7g     CALM1  0.11  STES
#> 44  hsa-let-7g   CASKIN1 -0.12  STES
#> 45  hsa-let-7g      CBX2 -0.12  STES
#> 46  hsa-let-7g      CCL7 -0.10  STES
#> 47  hsa-let-7g      CCNF -0.17  STES
#> 48  hsa-let-7g     CCNJL -0.11  STES
#> 49  hsa-let-7g     CD276 -0.16  STES
#> 50  hsa-let-7g    CDC14B  0.11  STES
#> 51  hsa-let-7g    CDC25A -0.11  STES
#> 52  hsa-let-7g  CDC42SE1 -0.23  STES
#> 53  hsa-let-7g     CDCA8 -0.13  STES
#> 54  hsa-let-7g    CERCAM -0.10  STES
#> 55  hsa-let-7g      CHRD  0.15  STES
#> 56  hsa-let-7g    CLASP2  0.14  STES
#> 57  hsa-let-7g      CLP1 -0.16  STES
#> 58  hsa-let-7g    CNOT6L  0.13  STES
#> 59  hsa-let-7g      COIL -0.11  STES
#> 60  hsa-let-7g   COL11A1 -0.12  STES
#> 61  hsa-let-7g   COL14A1  0.15  STES
#> 62  hsa-let-7g   COL15A1  0.19  STES
#> 63  hsa-let-7g    COL4A1  0.13  STES
#> 64  hsa-let-7g    COL4A2  0.15  STES
#> 65  hsa-let-7g    COL4A6  0.14  STES
#> 66  hsa-let-7g       CPD  0.14  STES
#> 67  hsa-let-7g     CPEB4  0.10  STES
#> 68  hsa-let-7g      CRBN  0.20  STES
#> 69  hsa-let-7g     CRTAP  0.13  STES
#> 70  hsa-let-7g      CRY2  0.15  STES
#> 71  hsa-let-7g      CTSC -0.17  STES
#> 72  hsa-let-7g   CXorf36  0.36  STES
#> 73  hsa-let-7g     CYTH3  0.14  STES
#> 74  hsa-let-7g     DAGLA  0.14  STES
#> 75  hsa-let-7g     DAPK1  0.21  STES
#> 76  hsa-let-7g    DCAF15 -0.21  STES
#> 77  hsa-let-7g     DCAF8  0.10  STES
#> 78  hsa-let-7g       DDN -0.19  STES
#> 79  hsa-let-7g      DLC1  0.26  STES
#> 80  hsa-let-7g       DMD  0.19  STES
#> 81  hsa-let-7g      DMP1 -0.10  STES
#> 82  hsa-let-7g      DNA2 -0.13  STES
#> 83  hsa-let-7g    DNAJA2 -0.11  STES
#> 84  hsa-let-7g    DNAJC1  0.17  STES
#> 85  hsa-let-7g     DNAL1  0.10  STES
#> 86  hsa-let-7g     DOCK3  0.24  STES
#> 87  hsa-let-7g      DPH3  0.15  STES
#> 88  hsa-let-7g    DPYSL3  0.18  STES
#> 89  hsa-let-7g      DSG3 -0.19  STES
#> 90  hsa-let-7g     DUSP7 -0.19  STES
#> 91  hsa-let-7g     DUSP9 -0.15  STES
#> 92  hsa-let-7g      DVL3 -0.15  STES
#> 93  hsa-let-7g    DYRK1A  0.11  STES
#> 94  hsa-let-7g     DYRK2  0.11  STES
#> 95  hsa-let-7g      E2F2 -0.13  STES
#> 96  hsa-let-7g       EDA  0.10  STES
#> 97  hsa-let-7g     EDEM3  0.11  STES
#> 98  hsa-let-7g      EDN1  0.24  STES
#> 99  hsa-let-7g      EEA1 -0.10  STES
#> 100 hsa-let-7g     EGLN2  0.11  STES
#> 101 hsa-let-7g    EIF4G2 -0.10  STES
#> 102 hsa-let-7g    ELOVL4 -0.12  STES
#> 103 hsa-let-7g     EPHA3  0.13  STES
#> 104 hsa-let-7g     ERCC4 -0.10  STES
#> 105 hsa-let-7g     ERCC6 -0.10  STES
#> 106 hsa-let-7g    ERGIC1  0.11  STES
#> 107 hsa-let-7g     ESPL1 -0.16  STES
#> 108 hsa-let-7g     ETNK2 -0.12  STES
#> 109 hsa-let-7g      EZH1  0.20  STES
#> 110 hsa-let-7g      EZH2 -0.13  STES
#> 111 hsa-let-7g  FAM103A1 -0.16  STES
#> 112 hsa-let-7g   FAM104A -0.12  STES
#> 113 hsa-let-7g   FAM118A  0.15  STES
#> 114 hsa-let-7g       FAS -0.10  STES
#> 115 hsa-let-7g    FBXL19 -0.20  STES
#> 116 hsa-let-7g    FBXO32  0.12  STES
#> 117 hsa-let-7g     FGF11 -0.20  STES
#> 118 hsa-let-7g    FNDC3A  0.14  STES
#> 119 hsa-let-7g     FOXN3  0.22  STES
#> 120 hsa-let-7g     FOXP1  0.24  STES
#> 121 hsa-let-7g     FOXP2  0.14  STES
#> 122 hsa-let-7g      FRS2  0.15  STES
#> 123 hsa-let-7g      FZD3 -0.10  STES
#> 124 hsa-let-7g      FZD4  0.19  STES
#> 125 hsa-let-7g      GAB2  0.15  STES
#> 126 hsa-let-7g      GALC  0.12  STES
#> 127 hsa-let-7g      GALE  0.10  STES
#> 128 hsa-let-7g    GALNT1 -0.15  STES
#> 129 hsa-let-7g    GALNT4  0.17  STES
#> 130 hsa-let-7g      GAS7  0.21  STES
#> 131 hsa-let-7g     GDPD1  0.17  STES
#> 132 hsa-let-7g    GEMIN7 -0.10  STES
#> 133 hsa-let-7g     GIPC1 -0.14  STES
#> 134 hsa-let-7g      GNAL  0.11  STES
#> 135 hsa-let-7g       GNS  0.12  STES
#> 136 hsa-let-7g    GOLGA7 -0.10  STES
#> 137 hsa-let-7g    GOLT1B -0.12  STES
#> 138 hsa-let-7g   GRID2IP  0.14  STES
#> 139 hsa-let-7g     GTF2I -0.13  STES
#> 140 hsa-let-7g     HABP4  0.12  STES
#> 141 hsa-let-7g     HCFC2  0.11  STES
#> 142 hsa-let-7g       HDX  0.12  STES
#> 143 hsa-let-7g      HIC1  0.11  STES
#> 144 hsa-let-7g       HK2 -0.16  STES
#> 145 hsa-let-7g       HLF  0.21  STES
#> 146 hsa-let-7g     HMGA1 -0.12  STES
#> 147 hsa-let-7g      HTR4  0.14  STES
#> 148 hsa-let-7g      IGF1  0.12  STES
#> 149 hsa-let-7g     IGF1R -0.18  STES
#> 150 hsa-let-7g    IKBKAP -0.10  STES
#> 151 hsa-let-7g      IL13  0.34  STES
#> 152 hsa-let-7g      IL6R  0.12  STES
#> 153 hsa-let-7g     IMPG2  0.21  STES
#> 154 hsa-let-7g    INPP5A  0.16  STES
#> 155 hsa-let-7g      INSR  0.24  STES
#> 156 hsa-let-7g     INTS2 -0.11  STES
#> 157 hsa-let-7g     ITGB3  0.24  STES
#> 158 hsa-let-7g     KCMF1 -0.20  STES
#> 159 hsa-let-7g     KCNC4  0.13  STES
#> 160 hsa-let-7g     KCNQ4  0.20  STES
#> 161 hsa-let-7g    KCTD10  0.21  STES
#> 162 hsa-let-7g     KIFC2 -0.14  STES
#> 163 hsa-let-7g      KLF9  0.27  STES
#> 164 hsa-let-7g    KLHL23  0.24  STES
#> 165 hsa-let-7g     KLHL6  0.10  STES
#> 166 hsa-let-7g   KREMEN1 -0.14  STES
#> 167 hsa-let-7g     LAMP2 -0.15  STES
#> 168 hsa-let-7g      LCOR  0.16  STES
#> 169 hsa-let-7g      LDB3  0.15  STES
#> 170 hsa-let-7g     LIMD1  0.13  STES
#> 171 hsa-let-7g     LIMK2 -0.19  STES
#> 172 hsa-let-7g      LIPH  0.13  STES
#> 173 hsa-let-7g      LMLN -0.14  STES
#> 174 hsa-let-7g     LOXL3  0.10  STES
#> 175 hsa-let-7g     LRIG1  0.12  STES
#> 176 hsa-let-7g   LRRC10B  0.13  STES
#> 177 hsa-let-7g     LSM11  0.38  STES
#> 178 hsa-let-7g    LUC7L3  0.15  STES
#> 179 hsa-let-7g    MAP3K3  0.23  STES
#> 180 hsa-let-7g    MAP4K3  0.15  STES
#> 181 hsa-let-7g    MAP4K4 -0.10  STES
#> 182 hsa-let-7g     MAPK6 -0.19  STES
#> 183 hsa-let-7g      MBD2 -0.26  STES
#> 184 hsa-let-7g      MDFI -0.11  STES
#> 185 hsa-let-7g      MED6 -0.15  STES
#> 186 hsa-let-7g     MEF2C  0.17  STES
#> 187 hsa-let-7g     MEF2D  0.20  STES
#> 188 hsa-let-7g     MEX3A -0.10  STES
#> 189 hsa-let-7g     MGAT3  0.10  STES
#> 190 hsa-let-7g    MGAT4A  0.13  STES
#> 191 hsa-let-7g      MGLL  0.15  STES
#> 192 hsa-let-7g     MLXIP  0.11  STES
#> 193 hsa-let-7g      MTDH -0.10  STES
#> 194 hsa-let-7g     MTUS1  0.12  STES
#> 195 hsa-let-7g     NCOA1  0.16  STES
#> 196 hsa-let-7g     NCOR1  0.30  STES
#> 197 hsa-let-7g      NEFM -0.11  STES
#> 198 hsa-let-7g      NEK3  0.10  STES
#> 199 hsa-let-7g    NHLRC2  0.12  STES
#> 200 hsa-let-7g    NHLRC3  0.17  STES
#> 201 hsa-let-7g    NIPAL4 -0.13  STES
#> 202 hsa-let-7g   NKIRAS2 -0.13  STES
#> 203 hsa-let-7g       NLN -0.13  STES
#> 204 hsa-let-7g     NOVA1  0.18  STES
#> 205 hsa-let-7g      NRAS -0.18  STES
#> 206 hsa-let-7g       NRK  0.10  STES
#> 207 hsa-let-7g      NTN1  0.26  STES
#> 208 hsa-let-7g     OLFM4  0.11  STES
#> 209 hsa-let-7g   ONECUT2  0.10  STES
#> 210 hsa-let-7g     OTUD3  0.10  STES
#> 211 hsa-let-7g     P2RX1  0.11  STES
#> 212 hsa-let-7g     PANX2 -0.11  STES
#> 213 hsa-let-7g    PARD6B  0.16  STES
#> 214 hsa-let-7g      PBX1  0.14  STES
#> 215 hsa-let-7g    PCYT1B -0.11  STES
#> 216 hsa-let-7g     PDE12  0.10  STES
#> 217 hsa-let-7g     PDGFB  0.21  STES
#> 218 hsa-let-7g      PDPR  0.10  STES
#> 219 hsa-let-7g   PHACTR2  0.17  STES
#> 220 hsa-let-7g      PHC3  0.14  STES
#> 221 hsa-let-7g      PIGA -0.10  STES
#> 222 hsa-let-7g    PLA2G3 -0.10  STES
#> 223 hsa-let-7g   PLEKHO1  0.10  STES
#> 224 hsa-let-7g    PLXNA4  0.12  STES
#> 225 hsa-let-7g    PLXNC1  0.14  STES
#> 226 hsa-let-7g    PLXND1  0.21  STES
#> 227 hsa-let-7g    PMAIP1 -0.24  STES
#> 228 hsa-let-7g      PNKD  0.12  STES
#> 229 hsa-let-7g    POLR2D -0.15  STES
#> 230 hsa-let-7g    POU2F1  0.12  STES
#> 231 hsa-let-7g  PPARGC1A  0.12  STES
#> 232 hsa-let-7g    PRKAA2  0.11  STES
#> 233 hsa-let-7g  PSORS1C2 -0.10  STES
#> 234 hsa-let-7g    PTP4A2  0.12  STES
#> 235 hsa-let-7g     PTPRU  0.10  STES
#> 236 hsa-let-7g     PYGO2 -0.11  STES
#> 237 hsa-let-7g     RAB8B  0.12  STES
#> 238 hsa-let-7g      RAG1 -0.14  STES
#> 239 hsa-let-7g   RALGPS1  0.13  STES
#> 240 hsa-let-7g    RANBP2  0.16  STES
#> 241 hsa-let-7g     RBMS2  0.21  STES
#> 242 hsa-let-7g    RIMKLA  0.13  STES
#> 243 hsa-let-7g    RNF165 -0.14  STES
#> 244 hsa-let-7g     RNF44 -0.12  STES
#> 245 hsa-let-7g      RNF7 -0.15  STES
#> 246 hsa-let-7g   RPS6KA3  0.12  STES
#> 247 hsa-let-7g      RRM2 -0.21  STES
#> 248 hsa-let-7g     RSPO2  0.11  STES
#> 249 hsa-let-7g      RTKN -0.15  STES
#> 250 hsa-let-7g   RUNX1T1  0.18  STES
#> 251 hsa-let-7g    S100A8 -0.10  STES
#> 252 hsa-let-7g    SAP30L  0.11  STES
#> 253 hsa-let-7g     SCN4B  0.11  STES
#> 254 hsa-let-7g   SEC14L1  0.18  STES
#> 255 hsa-let-7g    SEC31B  0.20  STES
#> 256 hsa-let-7g SECISBP2L  0.12  STES
#> 257 hsa-let-7g    SEMA4F -0.15  STES
#> 258 hsa-let-7g     SENP2 -0.13  STES
#> 259 hsa-let-7g     SENP5 -0.13  STES
#> 260 hsa-let-7g    SESTD1  0.24  STES
#> 261 hsa-let-7g    SFMBT1  0.15  STES
#> 262 hsa-let-7g      SGCD  0.14  STES
#> 263 hsa-let-7g     SH2B3  0.14  STES
#> 264 hsa-let-7g      SIM2  0.10  STES
#> 265 hsa-let-7g   SLC16A9 -0.15  STES
#> 266 hsa-let-7g  SLC22A23 -0.13  STES
#> 267 hsa-let-7g   SLC30A6 -0.10  STES
#> 268 hsa-let-7g   SLC30A7  0.10  STES
#> 269 hsa-let-7g   SLC31A2 -0.15  STES
#> 270 hsa-let-7g   SLC38A9 -0.12  STES
#> 271 hsa-let-7g   SLC45A4  0.11  STES
#> 272 hsa-let-7g   SLC46A3  0.20  STES
#> 273 hsa-let-7g   SLC6A15 -0.18  STES
#> 274 hsa-let-7g     SNX16 -0.13  STES
#> 275 hsa-let-7g   SPATA13  0.12  STES
#> 276 hsa-let-7g      SPEG  0.15  STES
#> 277 hsa-let-7g    SPRYD4 -0.12  STES
#> 278 hsa-let-7g    SPTBN4  0.15  STES
#> 279 hsa-let-7g   STARD13  0.38  STES
#> 280 hsa-let-7g     STK40  0.15  STES
#> 281 hsa-let-7g      STRN -0.13  STES
#> 282 hsa-let-7g   SYNJ2BP  0.16  STES
#> 283 hsa-let-7g     SYT11  0.19  STES
#> 284 hsa-let-7g    TARBP2 -0.14  STES
#> 285 hsa-let-7g      TBX5  0.16  STES
#> 286 hsa-let-7g      TET3 -0.14  STES
#> 287 hsa-let-7g    TEX261 -0.12  STES
#> 288 hsa-let-7g    TGFBR3  0.21  STES
#> 289 hsa-let-7g     THBS1  0.21  STES
#> 290 hsa-let-7g     THOC2  0.12  STES
#> 291 hsa-let-7g     TMEM2  0.12  STES
#> 292 hsa-let-7g   TMEM41A -0.12  STES
#> 293 hsa-let-7g    TMEM65 -0.13  STES
#> 294 hsa-let-7g     TMOD2  0.24  STES
#> 295 hsa-let-7g     TMTC3 -0.22  STES
#> 296 hsa-let-7g  TNFRSF1B  0.13  STES
#> 297 hsa-let-7g    TNFSF9 -0.11  STES
#> 298 hsa-let-7g      TP53 -0.10  STES
#> 299 hsa-let-7g      TPP1  0.11  STES
#> 300 hsa-let-7g   TRAPPC1 -0.19  STES
#> 301 hsa-let-7g   TSC22D2 -0.10  STES
#> 302 hsa-let-7g   TSPAN18  0.14  STES
#> 303 hsa-let-7g    TTC39C -0.10  STES
#> 304 hsa-let-7g       TTL -0.10  STES
#> 305 hsa-let-7g     UCHL5 -0.15  STES
#> 306 hsa-let-7g     UHRF1 -0.21  STES
#> 307 hsa-let-7g     USP12  0.28  STES
#> 308 hsa-let-7g     USP20  0.15  STES
#> 309 hsa-let-7g     USP21 -0.14  STES
#> 310 hsa-let-7g     USP38  0.10  STES
#> 311 hsa-let-7g     USP44  0.11  STES
#> 312 hsa-let-7g     USP47  0.13  STES
#> 313 hsa-let-7g      UTRN  0.16  STES
#> 314 hsa-let-7g    VANGL2 -0.16  STES
#> 315 hsa-let-7g      VAV3 -0.18  STES
#> 316 hsa-let-7g     VGLL3  0.13  STES
#> 317 hsa-let-7g     VSNL1 -0.16  STES
#> 318 hsa-let-7g     WIPI2 -0.11  STES
#> 319 hsa-let-7g     XYLT1  0.10  STES
#> 320 hsa-let-7g     YPEL2  0.28  STES
#> 321 hsa-let-7g     ZBTB5 -0.10  STES
#> 322 hsa-let-7g     ZC3H3 -0.16  STES
#> 323 hsa-let-7g    ZNF197  0.19  STES
#> 324 hsa-let-7g    ZNF200 -0.11  STES
#> 325 hsa-let-7g    ZNF282 -0.18  STES
#> 326 hsa-let-7g   ZNF512B -0.12  STES
#> 327 hsa-let-7g    ZNF644  0.10  STES
#> 328 hsa-let-7g    ZNF879 -0.11  STES
#> 329 hsa-let-7g    ZSWIM5  0.11  STES

# enter a custom query with different arguments
get_mir(conn,
        mir = 'hsa-let-7g',
        study = 'STES',
        min_abs_cor = .3,
        max_num = 5)
#>   mirna_base  feature  cor study
#> 1 hsa-let-7g ARHGEF15 0.39  STES
#> 2 hsa-let-7g    LSM11 0.38  STES
#> 3 hsa-let-7g  STARD13 0.38  STES
#> 4 hsa-let-7g  CXorf36 0.36  STES
#> 5 hsa-let-7g     IL13 0.34  STES