Get transcription factor correlations from cRegulome.db

This function access the sqlite database file which is obtained by running get_db. Basically, the function provides ways to query the database to the correlation data of the transcription factors of interest. The function returns an error if the database file cRegulome.db is not in the working directory.

Usage

get_tf(conn, tf, study, min_abs_cor, max_num, targets_only = FALSE, targets)

Arguments

conn

A connection such as this returned by dbConnect

tf

A required character vector of the transcription factor of interest. These are the HUGO official gene symbols of the genes contains the transcription factor.

study

A character vector of The Cancer Genome Atlas (TCGA) study identifiers. To view the available studies in TCGA project, https://tcga-data.nci.nih.gov/docs/publications/tcga. When left to default NULL all available studies will be included.

min_abs_cor

A numeric, an absolute correlation minimum between 0 and 1 for each mir.

max_num

An integer, maximum number of features to show for each mir in each study.

targets_only

A logical whether restrict the output to the recognized target features.

targets

A character vector of gene symbol names.

Value

A tidy data.frame of four columns. tf is the official gene symbols of the genes contains the transcription factor, feature

is the features/genes, cor is the corresponding expression correlations and study is TCGA study ID.

Examples

# locate the testset file and connect
fl <- system.file('extdata', 'cRegulome.db', package = 'cRegulome')
conn <- RSQLite::dbConnect(RSQLite::SQLite(), fl)

if (FALSE) {
# get transcription factors correlations in all studies
get_tf(conn,
        tf = 'LEF1')
}

# get correlations in a particular study
get_tf(conn,
       tf = 'LEF1',
       study = 'STES')
#>        tf    feature   cor study
#> 1    LEF1        A2M  0.55  STES
#> 2    LEF1     A4GALT  0.40  STES
#> 3    LEF1      ABCA1  0.44  STES
#> 4    LEF1      ABCA6  0.54  STES
#> 5    LEF1      ABCA8  0.41  STES
#> 6    LEF1      ABCA9  0.48  STES
#> 7    LEF1      ABCB1  0.56  STES
#> 8    LEF1      ABCB4  0.43  STES
#> 9    LEF1      ABCC9  0.54  STES
#> 10   LEF1       ABI3  0.45  STES
#> 11   LEF1     ABI3BP  0.50  STES
#> 12   LEF1      ACSL6  0.39  STES
#> 13   LEF1      ACSS3  0.52  STES
#> 14   LEF1      ACTA2  0.50  STES
#> 15   LEF1     ACVRL1  0.52  STES
#> 16   LEF1     ADAM19  0.56  STES
#> 17   LEF1     ADAM23  0.54  STES
#> 18   LEF1     ADAM33  0.51  STES
#> 19   LEF1   ADAMTS10  0.59  STES
#> 20   LEF1   ADAMTS12  0.39  STES
#> 21   LEF1   ADAMTS14  0.41  STES
#> 22   LEF1   ADAMTS18  0.47  STES
#> 23   LEF1    ADAMTS2  0.43  STES
#> 24   LEF1    ADAMTS3  0.52  STES
#> 25   LEF1    ADAMTS5  0.42  STES
#> 26   LEF1   ADAMTSL1  0.53  STES
#> 27   LEF1   ADAMTSL2  0.51  STES
#> 28   LEF1   ADAMTSL3  0.52  STES
#> 29   LEF1     ADARB1  0.46  STES
#> 30   LEF1        ADC  0.47  STES
#> 31   LEF1      ADCY2  0.54  STES
#> 32   LEF1      ADCY5  0.42  STES
#> 33   LEF1       ADD2  0.49  STES
#> 34   LEF1    ADORA2A  0.43  STES
#> 35   LEF1      ADPRH  0.53  STES
#> 36   LEF1      AEBP1  0.55  STES
#> 37   LEF1      AFAP1  0.39  STES
#> 38   LEF1       AFF3  0.47  STES
#> 39   LEF1      AKAP2  0.47  STES
#> 40   LEF1       AKT3  0.59  STES
#> 41   LEF1     AMOTL1  0.42  STES
#> 42   LEF1      AMPD3  0.39  STES
#> 43   LEF1       AMPH  0.51  STES
#> 44   LEF1     ANGPT1  0.44  STES
#> 45   LEF1    ANGPTL1  0.48  STES
#> 46   LEF1    ANGPTL2  0.48  STES
#> 47   LEF1       ANK2  0.58  STES
#> 48   LEF1       ANKH  0.51  STES
#> 49   LEF1    ANKRD44  0.43  STES
#> 50   LEF1    ANKRD55  0.42  STES
#> 51   LEF1     ANTXR1  0.55  STES
#> 52   LEF1      ANXA6  0.58  STES
#> 53   LEF1       AOAH  0.41  STES
#> 54   LEF1       AOC3  0.54  STES
#> 55   LEF1       AOX1  0.52  STES
#> 56   LEF1      AP1S2  0.51  STES
#> 57   LEF1      APBA2  0.47  STES
#> 58   LEF1      APBB1  0.49  STES
#> 59   LEF1      APLNR  0.43  STES
#> 60   LEF1       APOD  0.41  STES
#> 61   LEF1     APOLD1  0.40  STES
#> 62   LEF1         AR  0.52  STES
#> 63   LEF1   ARHGAP15  0.48  STES
#> 64   LEF1   ARHGAP20  0.51  STES
#> 65   LEF1   ARHGAP22  0.55  STES
#> 66   LEF1   ARHGAP25  0.46  STES
#> 67   LEF1   ARHGAP31  0.55  STES
#> 68   LEF1    ARHGDIB  0.39  STES
#> 69   LEF1   ARHGEF17  0.43  STES
#> 70   LEF1    ARHGEF6  0.52  STES
#> 71   LEF1     ARID5A  0.42  STES
#> 72   LEF1     ARID5B  0.42  STES
#> 73   LEF1      ARL10  0.51  STES
#> 74   LEF1      ARL4C  0.41  STES
#> 75   LEF1     ARMCX1  0.52  STES
#> 76   LEF1     ARMCX2  0.49  STES
#> 77   LEF1       ARSB  0.51  STES
#> 78   LEF1      ASAP1  0.38  STES
#> 79   LEF1       ASPN  0.52  STES
#> 80   LEF1      ASXL3  0.45  STES
#> 81   LEF1     ATP10A  0.50  STES
#> 82   LEF1     ATP1B2  0.54  STES
#> 83   LEF1   ATP6V1G2  0.52  STES
#> 84   LEF1     ATP8B2  0.57  STES
#> 85   LEF1     ATP8B4  0.43  STES
#> 86   LEF1      ATXN1  0.42  STES
#> 87   LEF1     AVPR1A  0.39  STES
#> 88   LEF1      BACE1  0.49  STES
#> 89   LEF1      BACH2  0.59  STES
#> 90   LEF1      BASP1  0.47  STES
#> 91   LEF1      BATF3  0.40  STES
#> 92   LEF1       BCL2  0.51  STES
#> 93   LEF1      BCL6B  0.44  STES
#> 94   LEF1      BEND5  0.45  STES
#> 95   LEF1      BEND6  0.53  STES
#> 96   LEF1       BEX4  0.44  STES
#> 97   LEF1        BGN  0.58  STES
#> 98   LEF1    BHLHE22  0.56  STES
#> 99   LEF1      BHMT2  0.51  STES
#> 100  LEF1      BICC1  0.55  STES
#> 101  LEF1       BIN2  0.39  STES
#> 102  LEF1       BMP6  0.41  STES
#> 103  LEF1      BMPER  0.45  STES
#> 104  LEF1       BNC2  0.58  STES
#> 105  LEF1        BOC  0.58  STES
#> 106  LEF1        BPI  0.41  STES
#> 107  LEF1        BTK  0.45  STES
#> 108  LEF1  C10orf105  0.38  STES
#> 109  LEF1  C10orf128  0.47  STES
#> 110  LEF1   C11orf63  0.40  STES
#> 111  LEF1   C14orf37  0.53  STES
#> 112  LEF1   C15orf59  0.50  STES
#> 113  LEF1   C16orf45  0.42  STES
#> 114  LEF1  C17orf103  0.45  STES
#> 115  LEF1   C19orf35  0.51  STES
#> 116  LEF1    C1QTNF2  0.53  STES
#> 117  LEF1    C1QTNF3  0.42  STES
#> 118  LEF1    C1QTNF7  0.47  STES
#> 119  LEF1        C1R  0.52  STES
#> 120  LEF1        C1S  0.53  STES
#> 121  LEF1   C21orf62  0.41  STES
#> 122  LEF1    C2orf40  0.42  STES
#> 123  LEF1         C7  0.46  STES
#> 124  LEF1    C8orf34  0.42  STES
#> 125  LEF1    C8orf48  0.46  STES
#> 126  LEF1    CACNA1C  0.46  STES
#> 127  LEF1    CACNA1H  0.51  STES
#> 128  LEF1      CADM1  0.48  STES
#> 129  LEF1      CADM3  0.46  STES
#> 130  LEF1     CALCRL  0.47  STES
#> 131  LEF1      CALD1  0.55  STES
#> 132  LEF1     CALHM2  0.55  STES
#> 133  LEF1     CAMK1D  0.54  STES
#> 134  LEF1      CAMK4  0.44  STES
#> 135  LEF1     CASP12  0.40  STES
#> 136  LEF1      CASQ2  0.46  STES
#> 137  LEF1   CCDC102A  0.42  STES
#> 138  LEF1   CCDC102B  0.41  STES
#> 139  LEF1      CCDC3  0.45  STES
#> 140  LEF1      CCDC8  0.54  STES
#> 141  LEF1     CCDC80  0.54  STES
#> 142  LEF1      CCL14  0.41  STES
#> 143  LEF1      CCL19  0.44  STES
#> 144  LEF1      CCL21  0.39  STES
#> 145  LEF1       CCR2  0.39  STES
#> 146  LEF1       CCR4  0.43  STES
#> 147  LEF1      CD180  0.43  STES
#> 148  LEF1       CD1C  0.42  STES
#> 149  LEF1       CD1D  0.51  STES
#> 150  LEF1       CD1E  0.41  STES
#> 151  LEF1      CD200  0.58  STES
#> 152  LEF1      CD226  0.38  STES
#> 153  LEF1       CD28  0.43  STES
#> 154  LEF1      CD302  0.42  STES
#> 155  LEF1       CD34  0.49  STES
#> 156  LEF1       CD36  0.38  STES
#> 157  LEF1       CD37  0.41  STES
#> 158  LEF1       CD48  0.39  STES
#> 159  LEF1        CD5  0.41  STES
#> 160  LEF1       CD53  0.39  STES
#> 161  LEF1      CD79B  0.40  STES
#> 162  LEF1       CD93  0.44  STES
#> 163  LEF1     CD99L2  0.48  STES
#> 164  LEF1   CDC42EP3  0.58  STES
#> 165  LEF1      CDH11  0.63  STES
#> 166  LEF1       CDH2  0.45  STES
#> 167  LEF1       CDH5  0.46  STES
#> 168  LEF1      CDK14  0.48  STES
#> 169  LEF1       CDO1  0.57  STES
#> 170  LEF1       CDR1  0.39  STES
#> 171  LEF1      CDYL2  0.42  STES
#> 172  LEF1   CEACAM21  0.40  STES
#> 173  LEF1      CECR1  0.41  STES
#> 174  LEF1      CELF2  0.51  STES
#> 175  LEF1      CERKL  0.46  STES
#> 176  LEF1       CES1  0.40  STES
#> 177  LEF1        CFH  0.44  STES
#> 178  LEF1      CGNL1  0.42  STES
#> 179  LEF1      CH25H  0.56  STES
#> 180  LEF1       CHN1  0.59  STES
#> 181  LEF1       CHRD  0.53  STES
#> 182  LEF1     CHRDL1  0.49  STES
#> 183  LEF1      CHST1  0.46  STES
#> 184  LEF1     CHST10  0.43  STES
#> 185  LEF1     CHST11  0.47  STES
#> 186  LEF1     CHST15  0.46  STES
#> 187  LEF1      CHST7  0.45  STES
#> 188  LEF1      CHST8  0.44  STES
#> 189  LEF1      CHSY3  0.51  STES
#> 190  LEF1       CILP  0.45  STES
#> 191  LEF1     CLDN11  0.49  STES
#> 192  LEF1    CLEC14A  0.48  STES
#> 193  LEF1     CLEC1A  0.45  STES
#> 194  LEF1     CLEC2B  0.42  STES
#> 195  LEF1     CLECL1  0.38  STES
#> 196  LEF1      CLIC2  0.57  STES
#> 197  LEF1      CLIP3  0.59  STES
#> 198  LEF1     CLSTN2  0.49  STES
#> 199  LEF1     CMKLR1  0.41  STES
#> 200  LEF1      CMTM3  0.54  STES
#> 201  LEF1      CMYA5  0.41  STES
#> 202  LEF1      CNIH3  0.47  STES
#> 203  LEF1       CNN3  0.44  STES
#> 204  LEF1     CNRIP1  0.59  STES
#> 205  LEF1      CNTN2  0.40  STES
#> 206  LEF1      CNTN4  0.58  STES
#> 207  LEF1    COL10A1  0.38  STES
#> 208  LEF1    COL12A1  0.39  STES
#> 209  LEF1    COL14A1  0.58  STES
#> 210  LEF1    COL15A1  0.43  STES
#> 211  LEF1    COL16A1  0.53  STES
#> 212  LEF1    COL18A1  0.47  STES
#> 213  LEF1    COL19A1  0.45  STES
#> 214  LEF1     COL1A2  0.50  STES
#> 215  LEF1     COL3A1  0.54  STES
#> 216  LEF1     COL4A1  0.38  STES
#> 217  LEF1     COL4A2  0.44  STES
#> 218  LEF1     COL4A3  0.38  STES
#> 219  LEF1     COL5A1  0.51  STES
#> 220  LEF1     COL5A2  0.48  STES
#> 221  LEF1     COL6A2  0.53  STES
#> 222  LEF1     COL6A3  0.54  STES
#> 223  LEF1     COL8A1  0.57  STES
#> 224  LEF1     COL8A2  0.54  STES
#> 225  LEF1    COLEC12  0.49  STES
#> 226  LEF1       COMP  0.46  STES
#> 227  LEF1      CORIN  0.48  STES
#> 228  LEF1     CORO2B  0.57  STES
#> 229  LEF1        CPE  0.53  STES
#> 230  LEF1      CPXM1  0.49  STES
#> 231  LEF1      CPXM2  0.52  STES
#> 232  LEF1        CPZ  0.51  STES
#> 233  LEF1        CR1  0.42  STES
#> 234  LEF1      CRHBP  0.43  STES
#> 235  LEF1   CRISPLD1  0.55  STES
#> 236  LEF1   CRISPLD2  0.53  STES
#> 237  LEF1      CRMP1  0.58  STES
#> 238  LEF1     CRTAC1  0.45  STES
#> 239  LEF1      CSF1R  0.38  STES
#> 240  LEF1     CSF2RB  0.44  STES
#> 241  LEF1      CSRP2  0.45  STES
#> 242  LEF1       CST2  0.43  STES
#> 243  LEF1       CTGF  0.44  STES
#> 244  LEF1     CTHRC1  0.48  STES
#> 245  LEF1       CTSK  0.56  STES
#> 246  LEF1       CTSO  0.43  STES
#> 247  LEF1       CUBN  0.47  STES
#> 248  LEF1     CX3CR1  0.43  STES
#> 249  LEF1     CXCL12  0.48  STES
#> 250  LEF1      CXCR4  0.42  STES
#> 251  LEF1    CXorf21  0.45  STES
#> 252  LEF1    CXorf36  0.48  STES
#> 253  LEF1     CYBRD1  0.54  STES
#> 254  LEF1       CYGB  0.57  STES
#> 255  LEF1     CYP1B1  0.52  STES
#> 256  LEF1    CYP27A1  0.48  STES
#> 257  LEF1     CYP2U1  0.48  STES
#> 258  LEF1     CYP7B1  0.52  STES
#> 259  LEF1    CYSLTR1  0.48  STES
#> 260  LEF1      CYYR1  0.52  STES
#> 261  LEF1      DAAM2  0.50  STES
#> 262  LEF1       DAB2  0.54  STES
#> 263  LEF1      DACT1  0.60  STES
#> 264  LEF1       DAZL  0.47  STES
#> 265  LEF1      DCHS1  0.58  STES
#> 266  LEF1      DCLK1  0.55  STES
#> 267  LEF1      DCLK2  0.56  STES
#> 268  LEF1        DCN  0.61  STES
#> 269  LEF1       DDR2  0.52  STES
#> 270  LEF1       DIO2  0.54  STES
#> 271  LEF1      DIRC3  0.40  STES
#> 272  LEF1      DISC1  0.54  STES
#> 273  LEF1       DKK2  0.45  STES
#> 274  LEF1       DKK3  0.53  STES
#> 275  LEF1       DLG4  0.51  STES
#> 276  LEF1       DLL1  0.39  STES
#> 277  LEF1        DMD  0.40  STES
#> 278  LEF1    DMRTC1B  0.38  STES
#> 279  LEF1    DNAJC18  0.53  STES
#> 280  LEF1     DOCK10  0.48  STES
#> 281  LEF1     DOCK11  0.54  STES
#> 282  LEF1      DOCK2  0.44  STES
#> 283  LEF1      DOCK8  0.40  STES
#> 284  LEF1       DOK5  0.56  STES
#> 285  LEF1       DOK6  0.64  STES
#> 286  LEF1        DPT  0.43  STES
#> 287  LEF1     DPYSL3  0.58  STES
#> 288  LEF1     DPYSL4  0.48  STES
#> 289  LEF1        DSE  0.41  STES
#> 290  LEF1       DSEL  0.55  STES
#> 291  LEF1       DTX1  0.47  STES
#> 292  LEF1     DUSP22  0.43  STES
#> 293  LEF1     DUSP26  0.41  STES
#> 294  LEF1       DYSF  0.39  STES
#> 295  LEF1      DZIP1  0.61  STES
#> 296  LEF1       EBF1  0.58  STES
#> 297  LEF1       EBF2  0.41  STES
#> 298  LEF1       EBF3  0.56  STES
#> 299  LEF1       ECM2  0.61  STES
#> 300  LEF1      EDA2R  0.48  STES
#> 301  LEF1      EDIL3  0.41  STES
#> 302  LEF1      EDNRA  0.58  STES
#> 303  LEF1     EFEMP1  0.55  STES
#> 304  LEF1     EFEMP2  0.55  STES
#> 305  LEF1      EFHD1  0.48  STES
#> 306  LEF1        EFS  0.52  STES
#> 307  LEF1       EGR2  0.50  STES
#> 308  LEF1       EHD3  0.54  STES
#> 309  LEF1       EID3  0.42  STES
#> 310  LEF1      ELMO1  0.59  STES
#> 311  LEF1        ELN  0.56  STES
#> 312  LEF1     ELOVL5  0.44  STES
#> 313  LEF1      ELTD1  0.44  STES
#> 314  LEF1       EMCN  0.48  STES
#> 315  LEF1       EML1  0.55  STES
#> 316  LEF1      ENOX1  0.68  STES
#> 317  LEF1      ENPEP  0.41  STES
#> 318  LEF1      ENPP2  0.46  STES
#> 319  LEF1     ENTPD1  0.60  STES
#> 320  LEF1    EPB41L3  0.48  STES
#> 321  LEF1      EPDR1  0.44  STES
#> 322  LEF1      EPHA3  0.59  STES
#> 323  LEF1      EPM2A  0.44  STES
#> 324  LEF1        ERG  0.56  STES
#> 325  LEF1       ESR1  0.50  STES
#> 326  LEF1        EVC  0.59  STES
#> 327  LEF1       EVC2  0.55  STES
#> 328  LEF1      EVI2A  0.44  STES
#> 329  LEF1      EVI2B  0.42  STES
#> 330  LEF1        EVL  0.50  STES
#> 331  LEF1        F10  0.39  STES
#> 332  LEF1        F2R  0.62  STES
#> 333  LEF1      F2RL2  0.47  STES
#> 334  LEF1         F8  0.46  STES
#> 335  LEF1      FAIM3  0.40  STES
#> 336  LEF1    FAM107A  0.43  STES
#> 337  LEF1    FAM110B  0.47  STES
#> 338  LEF1    FAM124A  0.44  STES
#> 339  LEF1    FAM124B  0.54  STES
#> 340  LEF1    FAM127C  0.46  STES
#> 341  LEF1     FAM13C  0.56  STES
#> 342  LEF1    FAM171B  0.43  STES
#> 343  LEF1    FAM180A  0.52  STES
#> 344  LEF1    FAM196B  0.40  STES
#> 345  LEF1    FAM198A  0.45  STES
#> 346  LEF1    FAM198B  0.53  STES
#> 347  LEF1    FAM19A5  0.56  STES
#> 348  LEF1     FAM20C  0.39  STES
#> 349  LEF1     FAM26E  0.53  STES
#> 350  LEF1     FAM43B  0.50  STES
#> 351  LEF1     FAM49A  0.46  STES
#> 352  LEF1     FAM65B  0.45  STES
#> 353  LEF1     FAM78A  0.42  STES
#> 354  LEF1        FAP  0.49  STES
#> 355  LEF1       FAT4  0.53  STES
#> 356  LEF1      FBLN1  0.54  STES
#> 357  LEF1      FBLN5  0.58  STES
#> 358  LEF1      FBLN7  0.57  STES
#> 359  LEF1       FBN1  0.59  STES
#> 360  LEF1      FBXL7  0.61  STES
#> 361  LEF1     FCGR2B  0.39  STES
#> 362  LEF1     FCGR2C  0.39  STES
#> 363  LEF1     FCHSD2  0.41  STES
#> 364  LEF1      FCRLA  0.38  STES
#> 365  LEF1     FERMT2  0.54  STES
#> 366  LEF1       FGD1  0.53  STES
#> 367  LEF1       FGD3  0.42  STES
#> 368  LEF1       FGD5  0.55  STES
#> 369  LEF1       FGF1  0.42  STES
#> 370  LEF1      FGF10  0.43  STES
#> 371  LEF1      FGF13  0.49  STES
#> 372  LEF1      FGF14  0.59  STES
#> 373  LEF1       FGF2  0.50  STES
#> 374  LEF1       FGF7  0.54  STES
#> 375  LEF1       FGL2  0.40  STES
#> 376  LEF1       FHL1  0.49  STES
#> 377  LEF1       FHL5  0.46  STES
#> 378  LEF1      FHOD3  0.51  STES
#> 379  LEF1      FIBIN  0.52  STES
#> 380  LEF1     FILIP1  0.48  STES
#> 381  LEF1    FILIP1L  0.56  STES
#> 382  LEF1       FLI1  0.55  STES
#> 383  LEF1      FLRT2  0.64  STES
#> 384  LEF1       FLT3  0.41  STES
#> 385  LEF1       FLT4  0.44  STES
#> 386  LEF1       FMO1  0.40  STES
#> 387  LEF1       FMO2  0.39  STES
#> 388  LEF1        FN1  0.41  STES
#> 389  LEF1      FNBP1  0.46  STES
#> 390  LEF1      FNDC1  0.55  STES
#> 391  LEF1      FOXC2  0.44  STES
#> 392  LEF1      FOXF1  0.53  STES
#> 393  LEF1      FOXI2  0.42  STES
#> 394  LEF1      FOXN3  0.41  STES
#> 395  LEF1      FOXP2  0.45  STES
#> 396  LEF1      FOXS1  0.58  STES
#> 397  LEF1      FREM1  0.48  STES
#> 398  LEF1     FRMD4A  0.45  STES
#> 399  LEF1      FRMD6  0.47  STES
#> 400  LEF1        FRY  0.43  STES
#> 401  LEF1       FRZB  0.55  STES
#> 402  LEF1      FSTL1  0.63  STES
#> 403  LEF1      FXYD1  0.46  STES
#> 404  LEF1      FXYD6  0.60  STES
#> 405  LEF1        FYB  0.41  STES
#> 406  LEF1        FYN  0.50  STES
#> 407  LEF1       FZD1  0.44  STES
#> 408  LEF1       GAB3  0.51  STES
#> 409  LEF1      GAP43  0.42  STES
#> 410  LEF1       GAPT  0.53  STES
#> 411  LEF1       GAS1  0.49  STES
#> 412  LEF1       GAS7  0.61  STES
#> 413  LEF1      GDF10  0.49  STES
#> 414  LEF1       GDF6  0.46  STES
#> 415  LEF1        GEM  0.45  STES
#> 416  LEF1      GFRA1  0.42  STES
#> 417  LEF1       GGT5  0.61  STES
#> 418  LEF1        GHR  0.52  STES
#> 419  LEF1     GIMAP1  0.48  STES
#> 420  LEF1     GIMAP4  0.48  STES
#> 421  LEF1     GIMAP5  0.41  STES
#> 422  LEF1     GIMAP6  0.49  STES
#> 423  LEF1     GIMAP7  0.46  STES
#> 424  LEF1     GIMAP8  0.50  STES
#> 425  LEF1       GJA4  0.51  STES
#> 426  LEF1       GJA5  0.44  STES
#> 427  LEF1       GLI1  0.65  STES
#> 428  LEF1       GLI2  0.39  STES
#> 429  LEF1       GLI3  0.56  STES
#> 430  LEF1     GLIPR1  0.55  STES
#> 431  LEF1      GLP2R  0.42  STES
#> 432  LEF1       GLRB  0.45  STES
#> 433  LEF1     GLT8D2  0.57  STES
#> 434  LEF1       GNAL  0.49  STES
#> 435  LEF1      GNAO1  0.52  STES
#> 436  LEF1       GNB4  0.52  STES
#> 437  LEF1      GNG11  0.49  STES
#> 438  LEF1       GNG2  0.55  STES
#> 439  LEF1      GNGT2  0.39  STES
#> 440  LEF1        GP5  0.48  STES
#> 441  LEF1       GPC6  0.55  STES
#> 442  LEF1      GPM6B  0.43  STES
#> 443  LEF1      GPNMB  0.40  STES
#> 444  LEF1     GPR116  0.39  STES
#> 445  LEF1     GPR124  0.65  STES
#> 446  LEF1     GPR133  0.51  STES
#> 447  LEF1     GPR155  0.49  STES
#> 448  LEF1     GPR162  0.55  STES
#> 449  LEF1     GPR173  0.40  STES
#> 450  LEF1     GPR176  0.43  STES
#> 451  LEF1     GPR183  0.43  STES
#> 452  LEF1      GPR21  0.45  STES
#> 453  LEF1      GPR34  0.44  STES
#> 454  LEF1      GPR65  0.39  STES
#> 455  LEF1      GPR68  0.46  STES
#> 456  LEF1      GPR88  0.44  STES
#> 457  LEF1    GPRASP1  0.58  STES
#> 458  LEF1    GPRASP2  0.46  STES
#> 459  LEF1     GPRC5A -0.39  STES
#> 460  LEF1       GPX7  0.51  STES
#> 461  LEF1       GRAP  0.52  STES
#> 462  LEF1      GRAPL  0.38  STES
#> 463  LEF1      GREB1  0.48  STES
#> 464  LEF1      GREM1  0.48  STES
#> 465  LEF1      GREM2  0.46  STES
#> 466  LEF1      GRID1  0.48  STES
#> 467  LEF1      GSPT2  0.38  STES
#> 468  LEF1      GSTM5  0.51  STES
#> 469  LEF1    GUCY1A2  0.41  STES
#> 470  LEF1    GUCY1A3  0.62  STES
#> 471  LEF1    GUCY1B3  0.59  STES
#> 472  LEF1     GXYLT2  0.56  STES
#> 473  LEF1       GYPC  0.52  STES
#> 474  LEF1      HAND2  0.43  STES
#> 475  LEF1       HAS2  0.41  STES
#> 476  LEF1      HCLS1  0.41  STES
#> 477  LEF1    HDGFRP3  0.51  STES
#> 478  LEF1        HDX  0.46  STES
#> 479  LEF1       HEY1  0.52  STES
#> 480  LEF1       HEY2  0.51  STES
#> 481  LEF1       HEYL  0.60  STES
#> 482  LEF1        HGF  0.46  STES
#> 483  LEF1       HHIP  0.38  STES
#> 484  LEF1       HIC1  0.59  STES
#> 485  LEF1      HMCN1  0.63  STES
#> 486  LEF1       HOPX  0.39  STES
#> 487  LEF1      HOXA4  0.42  STES
#> 488  LEF1      HOXD8  0.40  STES
#> 489  LEF1       HRH2  0.40  STES
#> 490  LEF1   HS3ST3A1  0.43  STES
#> 491  LEF1   HS3ST3B1  0.40  STES
#> 492  LEF1    HSD11B1  0.46  STES
#> 493  LEF1      HSPB2  0.53  STES
#> 494  LEF1      HSPB7  0.42  STES
#> 495  LEF1      HTR1F  0.39  STES
#> 496  LEF1      HTR2A  0.56  STES
#> 497  LEF1      HTRA3  0.50  STES
#> 498  LEF1        ID4  0.46  STES
#> 499  LEF1        IDS  0.41  STES
#> 500  LEF1      IFFO1  0.58  STES
#> 501  LEF1     IGDCC4  0.52  STES
#> 502  LEF1       IGF1  0.56  STES
#> 503  LEF1       IGF2  0.48  STES
#> 504  LEF1     IGFBP3  0.48  STES
#> 505  LEF1     IGFBP5  0.59  STES
#> 506  LEF1     IGFBP7  0.61  STES
#> 507  LEF1     IGSF10  0.42  STES
#> 508  LEF1      IKZF1  0.42  STES
#> 509  LEF1     IL10RA  0.41  STES
#> 510  LEF1      IL1R1  0.52  STES
#> 511  LEF1       IL33  0.39  STES
#> 512  LEF1       IL7R  0.41  STES
#> 513  LEF1      INHBA  0.45  STES
#> 514  LEF1       INMT  0.58  STES
#> 515  LEF1     IQSEC3  0.43  STES
#> 516  LEF1       ISLR  0.55  STES
#> 517  LEF1      ISLR2  0.54  STES
#> 518  LEF1       ISM1  0.45  STES
#> 519  LEF1      ITGA1  0.47  STES
#> 520  LEF1     ITGA11  0.49  STES
#> 521  LEF1      ITGA4  0.48  STES
#> 522  LEF1      ITGA7  0.49  STES
#> 523  LEF1      ITGA9  0.52  STES
#> 524  LEF1      ITIH3  0.50  STES
#> 525  LEF1      ITIH5  0.51  STES
#> 526  LEF1      ITPKB  0.44  STES
#> 527  LEF1      ITPR1  0.47  STES
#> 528  LEF1     ITPRIP  0.40  STES
#> 529  LEF1       JAM2  0.53  STES
#> 530  LEF1       JAM3  0.60  STES
#> 531  LEF1      JAZF1  0.41  STES
#> 532  LEF1       JPH4  0.48  STES
#> 533  LEF1       KAL1  0.47  STES
#> 534  LEF1    KATNAL1  0.52  STES
#> 535  LEF1      KCNA2  0.43  STES
#> 536  LEF1      KCNA6  0.41  STES
#> 537  LEF1      KCND1  0.48  STES
#> 538  LEF1      KCNE4  0.54  STES
#> 539  LEF1     KCNIP1  0.43  STES
#> 540  LEF1      KCNJ5  0.38  STES
#> 541  LEF1      KCNJ8  0.51  STES
#> 542  LEF1     KCNK12  0.46  STES
#> 543  LEF1     KCNK13  0.42  STES
#> 544  LEF1      KCNK2  0.39  STES
#> 545  LEF1     KCNMB1  0.41  STES
#> 546  LEF1      KCNN3  0.44  STES
#> 547  LEF1      KCNQ5  0.41  STES
#> 548  LEF1      KCNT2  0.54  STES
#> 549  LEF1     KCTD12  0.42  STES
#> 550  LEF1     KDELC1  0.39  STES
#> 551  LEF1       KERA  0.50  STES
#> 552  LEF1   KIAA0355  0.50  STES
#> 553  LEF1   KIAA0922  0.41  STES
#> 554  LEF1   KIAA1462  0.46  STES
#> 555  LEF1   KIAA1614  0.54  STES
#> 556  LEF1   KIAA1755  0.67  STES
#> 557  LEF1      KIF3C  0.45  STES
#> 558  LEF1     KIRREL  0.54  STES
#> 559  LEF1        KIT  0.42  STES
#> 560  LEF1      KLF12  0.54  STES
#> 561  LEF1       KLF8  0.44  STES
#> 562  LEF1     KLHDC1  0.39  STES
#> 563  LEF1     KLHL14  0.39  STES
#> 564  LEF1     KLHL38  0.43  STES
#> 565  LEF1      KLHL5  0.45  STES
#> 566  LEF1      KLHL6  0.42  STES
#> 567  LEF1      KLRG1  0.43  STES
#> 568  LEF1      LAIR1  0.41  STES
#> 569  LEF1      LAMA2  0.60  STES
#> 570  LEF1      LAMA4  0.57  STES
#> 571  LEF1      LARP6  0.39  STES
#> 572  LEF1       LAYN  0.53  STES
#> 573  LEF1        LBH  0.55  STES
#> 574  LEF1       LCP2  0.38  STES
#> 575  LEF1       LDB2  0.56  STES
#> 576  LEF1      LDOC1  0.49  STES
#> 577  LEF1       LGI2  0.62  STES
#> 578  LEF1       LGI4  0.47  STES
#> 579  LEF1       LHFP  0.63  STES
#> 580  LEF1     LILRA1  0.39  STES
#> 581  LEF1     LILRA4  0.43  STES
#> 582  LEF1     LILRB1  0.40  STES
#> 583  LEF1      LIMS2  0.50  STES
#> 584  LEF1     LINGO3  0.46  STES
#> 585  LEF1      LIX1L  0.52  STES
#> 586  LEF1       LMO2  0.48  STES
#> 587  LEF1       LMO3  0.51  STES
#> 588  LEF1  LOC283070  0.57  STES
#> 589  LEF1  LOC283731  0.51  STES
#> 590  LEF1  LOC401463  0.47  STES
#> 591  LEF1  LOC441204  0.50  STES
#> 592  LEF1  LOC653653  0.44  STES
#> 593  LEF1        LOX  0.38  STES
#> 594  LEF1      LPAR4  0.53  STES
#> 595  LEF1      LPPR4  0.54  STES
#> 596  LEF1       LPXN  0.40  STES
#> 597  LEF1      LRCH2  0.56  STES
#> 598  LEF1      LRFN5  0.44  STES
#> 599  LEF1      LRP12  0.44  STES
#> 600  LEF1     LRRC15  0.38  STES
#> 601  LEF1     LRRC17  0.50  STES
#> 602  LEF1     LRRC18  0.45  STES
#> 603  LEF1      LRRC2  0.40  STES
#> 604  LEF1     LRRC32  0.63  STES
#> 605  LEF1     LRRC4C  0.55  STES
#> 606  LEF1     LRRC55  0.47  STES
#> 607  LEF1     LRRC70  0.45  STES
#> 608  LEF1     LRRC8C  0.47  STES
#> 609  LEF1      LRRK2  0.44  STES
#> 610  LEF1      LRRN2  0.42  STES
#> 611  LEF1      LRRN3  0.42  STES
#> 612  LEF1      LSAMP  0.63  STES
#> 613  LEF1      LTBP1  0.52  STES
#> 614  LEF1      LTBP2  0.60  STES
#> 615  LEF1        LUM  0.59  STES
#> 616  LEF1       LY96  0.45  STES
#> 617  LEF1      LZTS1  0.55  STES
#> 618  LEF1        MAF  0.59  STES
#> 619  LEF1       MAFB  0.48  STES
#> 620  LEF1    MAGED4B  0.38  STES
#> 621  LEF1     MAGEE1  0.42  STES
#> 622  LEF1     MAGEH1  0.49  STES
#> 623  LEF1     MAGEL2  0.53  STES
#> 624  LEF1      MAGI2  0.45  STES
#> 625  LEF1     MAMDC2  0.40  STES
#> 626  LEF1     MAN1C1  0.53  STES
#> 627  LEF1      MAP1A  0.62  STES
#> 628  LEF1      MAP1B  0.54  STES
#> 629  LEF1     MAP3K3  0.50  STES
#> 630  LEF1       MAP6  0.52  STES
#> 631  LEF1     MAP7D3  0.43  STES
#> 632  LEF1     MAPK10  0.58  STES
#> 633  LEF1     MARCH1  0.40  STES
#> 634  LEF1        MCC  0.56  STES
#> 635  LEF1      MCTP1  0.41  STES
#> 636  LEF1      MDFIC  0.55  STES
#> 637  LEF1      MEF2C  0.59  STES
#> 638  LEF1       MEG3  0.43  STES
#> 639  LEF1      MEIS3  0.57  STES
#> 640  LEF1      MEOX1  0.42  STES
#> 641  LEF1      MEOX2  0.51  STES
#> 642  LEF1      MEX3B  0.57  STES
#> 643  LEF1      MFAP5  0.51  STES
#> 644  LEF1      MFGE8  0.49  STES
#> 645  LEF1       MFNG  0.44  STES
#> 646  LEF1     MGAT5B  0.41  STES
#> 647  LEF1        MGP  0.57  STES
#> 648  LEF1       MID2  0.42  STES
#> 649  LEF1        MKX  0.43  STES
#> 650  LEF1     MLLT11  0.39  STES
#> 651  LEF1      MMP16  0.52  STES
#> 652  LEF1      MMP17  0.40  STES
#> 653  LEF1       MMP2  0.50  STES
#> 654  LEF1     MMP23B  0.45  STES
#> 655  LEF1      MMRN1  0.40  STES
#> 656  LEF1      MMRN2  0.51  STES
#> 657  LEF1        MN1  0.57  STES
#> 658  LEF1      MOCS1  0.42  STES
#> 659  LEF1      MOXD1  0.59  STES
#> 660  LEF1       MPDZ  0.56  STES
#> 661  LEF1      MPEG1  0.41  STES
#> 662  LEF1       MPP1  0.43  STES
#> 663  LEF1       MRAS  0.53  STES
#> 664  LEF1       MRC2  0.53  STES
#> 665  LEF1     MRGPRF  0.49  STES
#> 666  LEF1      MRVI1  0.54  STES
#> 667  LEF1     MS4A6A  0.40  STES
#> 668  LEF1      MS4A7  0.41  STES
#> 669  LEF1        MSC  0.54  STES
#> 670  LEF1        MSN  0.49  STES
#> 671  LEF1      MSRB3  0.56  STES
#> 672  LEF1       MUSK  0.39  STES
#> 673  LEF1      MYCT1  0.51  STES
#> 674  LEF1      MYH10  0.48  STES
#> 675  LEF1       MYLK  0.53  STES
#> 676  LEF1      MYO5A  0.47  STES
#> 677  LEF1      MYOCD  0.41  STES
#> 678  LEF1      MYOZ3  0.51  STES
#> 679  LEF1      NALCN  0.58  STES
#> 680  LEF1     NAP1L2  0.40  STES
#> 681  LEF1     NAP1L3  0.61  STES
#> 682  LEF1       NAV3  0.54  STES
#> 683  LEF1      NCALD  0.43  STES
#> 684  LEF1      NCAM1  0.49  STES
#> 685  LEF1    NCKAP1L  0.41  STES
#> 686  LEF1        NDN  0.57  STES
#> 687  LEF1      NEGR1  0.50  STES
#> 688  LEF1       NEXN  0.44  STES
#> 689  LEF1      NFASC  0.53  STES
#> 690  LEF1        NGF  0.42  STES
#> 691  LEF1       NGFR  0.51  STES
#> 692  LEF1      NHSL2  0.54  STES
#> 693  LEF1       NID1  0.49  STES
#> 694  LEF1  NIPSNAP3B  0.48  STES
#> 695  LEF1      NLGN3  0.42  STES
#> 696  LEF1      NLRP3  0.39  STES
#> 697  LEF1        NOG  0.41  STES
#> 698  LEF1     NOTCH3  0.43  STES
#> 699  LEF1      NOVA1  0.46  STES
#> 700  LEF1       NOX4  0.52  STES
#> 701  LEF1      NPAS3  0.51  STES
#> 702  LEF1      NR2F1  0.54  STES
#> 703  LEF1      NR3C1  0.43  STES
#> 704  LEF1      NRIP2  0.53  STES
#> 705  LEF1        NRK  0.50  STES
#> 706  LEF1       NRP1  0.42  STES
#> 707  LEF1       NRP2  0.50  STES
#> 708  LEF1      NRXN2  0.56  STES
#> 709  LEF1        NTM  0.45  STES
#> 710  LEF1      NTNG2  0.56  STES
#> 711  LEF1      NTRK3  0.50  STES
#> 712  LEF1      NUAK1  0.52  STES
#> 713  LEF1     NUDT10  0.50  STES
#> 714  LEF1     NUDT11  0.50  STES
#> 715  LEF1      NXPH3  0.51  STES
#> 716  LEF1        OGN  0.55  STES
#> 717  LEF1      OLFM1  0.40  STES
#> 718  LEF1     OLFML1  0.66  STES
#> 719  LEF1    OLFML2B  0.53  STES
#> 720  LEF1     OLFML3  0.52  STES
#> 721  LEF1        OMD  0.56  STES
#> 722  LEF1     OSBPL6  0.42  STES
#> 723  LEF1      OSTM1  0.41  STES
#> 724  LEF1      P2RX7  0.46  STES
#> 725  LEF1     P2RY12  0.42  STES
#> 726  LEF1     P2RY13  0.39  STES
#> 727  LEF1     P2RY14  0.43  STES
#> 728  LEF1      P4HA3  0.47  STES
#> 729  LEF1    PABPC4L  0.55  STES
#> 730  LEF1     PABPC5  0.61  STES
#> 731  LEF1      PALLD  0.44  STES
#> 732  LEF1      PALM2  0.40  STES
#> 733  LEF1      PALMD  0.44  STES
#> 734  LEF1      PAPLN  0.53  STES
#> 735  LEF1     PARD6G  0.41  STES
#> 736  LEF1      PARVA  0.39  STES
#> 737  LEF1      PARVG  0.45  STES
#> 738  LEF1      PCBP3  0.51  STES
#> 739  LEF1     PCDH12  0.41  STES
#> 740  LEF1     PCDH17  0.46  STES
#> 741  LEF1     PCDH18  0.51  STES
#> 742  LEF1     PCDH19  0.51  STES
#> 743  LEF1    PCDHB18  0.44  STES
#> 744  LEF1     PCDHB4  0.49  STES
#> 745  LEF1     PCDHB7  0.41  STES
#> 746  LEF1   PCDHGA12  0.54  STES
#> 747  LEF1    PCDHGA2  0.46  STES
#> 748  LEF1    PCDHGB6  0.49  STES
#> 749  LEF1    PCDHGB7  0.55  STES
#> 750  LEF1      PDE1A  0.62  STES
#> 751  LEF1      PDE1B  0.44  STES
#> 752  LEF1      PDE2A  0.41  STES
#> 753  LEF1      PDE3A  0.48  STES
#> 754  LEF1      PDE4B  0.40  STES
#> 755  LEF1      PDE7B  0.44  STES
#> 756  LEF1      PDE8B  0.43  STES
#> 757  LEF1      PDGFC  0.42  STES
#> 758  LEF1     PDGFRA  0.45  STES
#> 759  LEF1     PDGFRB  0.61  STES
#> 760  LEF1     PDLIM3  0.54  STES
#> 761  LEF1     PDLIM4  0.41  STES
#> 762  LEF1       PDPN  0.44  STES
#> 763  LEF1      PDZD4  0.51  STES
#> 764  LEF1     PDZRN3  0.54  STES
#> 765  LEF1     PECAM1  0.48  STES
#> 766  LEF1       PEG3  0.57  STES
#> 767  LEF1       PGM5  0.47  STES
#> 768  LEF1        PGR  0.54  STES
#> 769  LEF1    PHACTR1  0.49  STES
#> 770  LEF1       PHC1  0.41  STES
#> 771  LEF1     PHF21A  0.45  STES
#> 772  LEF1     PHLDB1  0.55  STES
#> 773  LEF1     PHLDB2  0.38  STES
#> 774  LEF1    PHYHIPL  0.57  STES
#> 775  LEF1       PID1  0.43  STES
#> 776  LEF1     PIK3R5  0.40  STES
#> 777  LEF1     PIK3R6  0.47  STES
#> 778  LEF1       PKD2  0.53  STES
#> 779  LEF1       PKIA  0.41  STES
#> 780  LEF1       PKIG  0.51  STES
#> 781  LEF1     PKNOX2  0.47  STES
#> 782  LEF1    PLA2G4C  0.53  STES
#> 783  LEF1     PLA2G5  0.42  STES
#> 784  LEF1      PLAC9  0.53  STES
#> 785  LEF1     PLAGL1  0.40  STES
#> 786  LEF1      PLCG2  0.46  STES
#> 787  LEF1      PLCL1  0.59  STES
#> 788  LEF1    PLEKHO1  0.47  STES
#> 789  LEF1    PLEKHO2  0.51  STES
#> 790  LEF1        PLN  0.46  STES
#> 791  LEF1     PLSCR4  0.44  STES
#> 792  LEF1     PLXDC1  0.58  STES
#> 793  LEF1     PLXDC2  0.58  STES
#> 794  LEF1     PLXNA4  0.53  STES
#> 795  LEF1     PLXNC1  0.45  STES
#> 796  LEF1      PMP22  0.49  STES
#> 797  LEF1     PNMAL1  0.51  STES
#> 798  LEF1     PNMAL2  0.58  STES
#> 799  LEF1       PODN  0.56  STES
#> 800  LEF1     POPDC2  0.47  STES
#> 801  LEF1      POSTN  0.44  STES
#> 802  LEF1     POU2F2  0.49  STES
#> 803  LEF1     POU6F1  0.49  STES
#> 804  LEF1     PPAP2A  0.44  STES
#> 805  LEF1     PPAP2B  0.50  STES
#> 806  LEF1    PPAPDC3  0.52  STES
#> 807  LEF1     PPFIA2  0.51  STES
#> 808  LEF1      PPM1K  0.41  STES
#> 809  LEF1    PPP2R2B  0.42  STES
#> 810  LEF1      PRDM6  0.54  STES
#> 811  LEF1      PRELP  0.56  STES
#> 812  LEF1      PREX1  0.45  STES
#> 813  LEF1      PREX2  0.46  STES
#> 814  LEF1   PRICKLE1  0.50  STES
#> 815  LEF1   PRICKLE2  0.52  STES
#> 816  LEF1    PRKAR2B  0.47  STES
#> 817  LEF1      PRKCB  0.42  STES
#> 818  LEF1      PRKCH  0.41  STES
#> 819  LEF1      PRKD1  0.53  STES
#> 820  LEF1      PRKG1  0.53  STES
#> 821  LEF1       PRND  0.44  STES
#> 822  LEF1      PRPH2  0.46  STES
#> 823  LEF1      PRR16  0.42  STES
#> 824  LEF1      PRRX1  0.54  STES
#> 825  LEF1     PRSS23  0.45  STES
#> 826  LEF1     PRUNE2  0.45  STES
#> 827  LEF1      PTCH2  0.56  STES
#> 828  LEF1      PTGDR  0.48  STES
#> 829  LEF1     PTGER3  0.57  STES
#> 830  LEF1      PTGFR  0.55  STES
#> 831  LEF1      PTGIR  0.52  STES
#> 832  LEF1      PTGIS  0.51  STES
#> 833  LEF1      PTH1R  0.46  STES
#> 834  LEF1        PTN  0.43  STES
#> 835  LEF1     PTP4A3  0.43  STES
#> 836  LEF1    PTPLAD2  0.51  STES
#> 837  LEF1      PTPRB  0.40  STES
#> 838  LEF1      PTPRC  0.39  STES
#> 839  LEF1      PTPRD  0.51  STES
#> 840  LEF1      PTPRM  0.52  STES
#> 841  LEF1      PTPRO  0.43  STES
#> 842  LEF1       PTRF  0.49  STES
#> 843  LEF1       PURG  0.48  STES
#> 844  LEF1       PXDN  0.41  STES
#> 845  LEF1      RAB23  0.42  STES
#> 846  LEF1      RAB30  0.51  STES
#> 847  LEF1      RAB31  0.50  STES
#> 848  LEF1     RAB33A  0.40  STES
#> 849  LEF1     RAB39B  0.42  STES
#> 850  LEF1      RAI14  0.41  STES
#> 851  LEF1       RAI2  0.52  STES
#> 852  LEF1      RAMP3  0.48  STES
#> 853  LEF1      RASD2  0.45  STES
#> 854  LEF1    RASGRF2  0.55  STES
#> 855  LEF1    RASGRP3  0.43  STES
#> 856  LEF1     RASL12  0.53  STES
#> 857  LEF1     RASSF2  0.53  STES
#> 858  LEF1     RASSF8  0.49  STES
#> 859  LEF1      RBMS3  0.61  STES
#> 860  LEF1       RBP1  0.46  STES
#> 861  LEF1      RCAN2  0.54  STES
#> 862  LEF1     RCBTB2  0.48  STES
#> 863  LEF1      RCSD1  0.50  STES
#> 864  LEF1      RCVRN  0.40  STES
#> 865  LEF1       REM1  0.52  STES
#> 866  LEF1       RERG  0.47  STES
#> 867  LEF1      RFTN1  0.66  STES
#> 868  LEF1      RFTN2  0.58  STES
#> 869  LEF1      RGAG4  0.59  STES
#> 870  LEF1       RGL1  0.50  STES
#> 871  LEF1      RGS18  0.41  STES
#> 872  LEF1      RGS22  0.42  STES
#> 873  LEF1       RGS4  0.45  STES
#> 874  LEF1       RGS5  0.39  STES
#> 875  LEF1     RGS7BP  0.53  STES
#> 876  LEF1       RHOJ  0.56  STES
#> 877  LEF1       RHOQ  0.46  STES
#> 878  LEF1     RNASE6  0.43  STES
#> 879  LEF1     RNF130  0.41  STES
#> 880  LEF1    RNF144A  0.45  STES
#> 881  LEF1     RNF150  0.57  STES
#> 882  LEF1     RNF180  0.55  STES
#> 883  LEF1     RNF217  0.44  STES
#> 884  LEF1      ROBO1  0.52  STES
#> 885  LEF1       ROR2  0.47  STES
#> 886  LEF1       RORA  0.50  STES
#> 887  LEF1       RORB  0.46  STES
#> 888  LEF1    RPLP0P2  0.40  STES
#> 889  LEF1      RSPO1  0.45  STES
#> 890  LEF1      RSPO3  0.51  STES
#> 891  LEF1       RTN1  0.49  STES
#> 892  LEF1    RUNDC3B  0.41  STES
#> 893  LEF1    RUNX1T1  0.64  STES
#> 894  LEF1      S1PR1  0.52  STES
#> 895  LEF1      S1PR2  0.43  STES
#> 896  LEF1      S1PR3  0.61  STES
#> 897  LEF1      SALL2  0.51  STES
#> 898  LEF1      SARDH  0.43  STES
#> 899  LEF1      SASH3  0.41  STES
#> 900  LEF1     SCARF2  0.54  STES
#> 901  LEF1       SCG5  0.46  STES
#> 902  LEF1     SCHIP1  0.39  STES
#> 903  LEF1      SCN1B  0.47  STES
#> 904  LEF1      SCN2B  0.48  STES
#> 905  LEF1      SCN3B  0.44  STES
#> 906  LEF1      SCN4B  0.49  STES
#> 907  LEF1      SCRG1  0.39  STES
#> 908  LEF1     SCUBE2  0.48  STES
#> 909  LEF1       SDC2  0.60  STES
#> 910  LEF1       SDPR  0.49  STES
#> 911  LEF1       SELL  0.40  STES
#> 912  LEF1       SELP  0.46  STES
#> 913  LEF1     SEMA5A  0.40  STES
#> 914  LEF1     SEMA5B  0.40  STES
#> 915  LEF1      SEPT5  0.53  STES
#> 916  LEF1   SERPINE2  0.56  STES
#> 917  LEF1   SERPINF1  0.59  STES
#> 918  LEF1   SERPINI1  0.40  STES
#> 919  LEF1      SESN3  0.41  STES
#> 920  LEF1     SETBP1  0.54  STES
#> 921  LEF1     SFMBT2  0.46  STES
#> 922  LEF1      SFRP1  0.46  STES
#> 923  LEF1      SFRP2  0.45  STES
#> 924  LEF1      SFRP4  0.55  STES
#> 925  LEF1       SGCD  0.56  STES
#> 926  LEF1       SGCE  0.45  STES
#> 927  LEF1      SGIP1  0.54  STES
#> 928  LEF1       SGTB  0.48  STES
#> 929  LEF1    SH3BGRL  0.40  STES
#> 930  LEF1     SH3BP5  0.54  STES
#> 931  LEF1   SH3PXD2B  0.45  STES
#> 932  LEF1     SH3RF3  0.59  STES
#> 933  LEF1       SHC4  0.51  STES
#> 934  LEF1        SHE  0.53  STES
#> 935  LEF1     SHISA2  0.50  STES
#> 936  LEF1     SHISA3  0.49  STES
#> 937  LEF1    SHROOM4  0.46  STES
#> 938  LEF1    SIGLEC6  0.41  STES
#> 939  LEF1     SLAMF1  0.41  STES
#> 940  LEF1    SLC24A3  0.55  STES
#> 941  LEF1    SLC24A4  0.41  STES
#> 942  LEF1    SLC35F1  0.46  STES
#> 943  LEF1    SLC45A1  0.54  STES
#> 944  LEF1     SLC8A1  0.51  STES
#> 945  LEF1     SLC9A9  0.51  STES
#> 946  LEF1    SLCO1C1  0.39  STES
#> 947  LEF1      SLIT2  0.58  STES
#> 948  LEF1      SLIT3  0.54  STES
#> 949  LEF1      SMAD9  0.51  STES
#> 950  LEF1      SMOC2  0.55  STES
#> 951  LEF1      SNAI2  0.48  STES
#> 952  LEF1     SNAP91  0.40  STES
#> 953  LEF1      SNRPN  0.50  STES
#> 954  LEF1       SOBP  0.47  STES
#> 955  LEF1       SOD3  0.44  STES
#> 956  LEF1     SORCS1  0.46  STES
#> 957  LEF1     SORCS2  0.55  STES
#> 958  LEF1      SOX10  0.41  STES
#> 959  LEF1      SOX17  0.48  STES
#> 960  LEF1       SOX5  0.56  STES
#> 961  LEF1      SPARC  0.52  STES
#> 962  LEF1    SPARCL1  0.52  STES
#> 963  LEF1      SPG20  0.46  STES
#> 964  LEF1     SPOCK1  0.44  STES
#> 965  LEF1      SPON1  0.50  STES
#> 966  LEF1      SPON2  0.48  STES
#> 967  LEF1      SPSB1  0.44  STES
#> 968  LEF1       SRPX  0.43  STES
#> 969  LEF1      SRPX2  0.47  STES
#> 970  LEF1      SSC5D  0.55  STES
#> 971  LEF1       SSPN  0.51  STES
#> 972  LEF1      SSTR2  0.52  STES
#> 973  LEF1    ST3GAL3  0.44  STES
#> 974  LEF1    ST6GAL2  0.51  STES
#> 975  LEF1 ST6GALNAC5  0.49  STES
#> 976  LEF1    ST8SIA1  0.53  STES
#> 977  LEF1    ST8SIA4  0.50  STES
#> 978  LEF1      STAP1  0.38  STES
#> 979  LEF1     STARD8  0.44  STES
#> 980  LEF1     STK32B  0.39  STES
#> 981  LEF1      STMN2  0.49  STES
#> 982  LEF1      STON1  0.57  STES
#> 983  LEF1      STRA6  0.42  STES
#> 984  LEF1      SULF1  0.49  STES
#> 985  LEF1    SULT1C4  0.59  STES
#> 986  LEF1      SUSD5  0.44  STES
#> 987  LEF1       SV2B  0.56  STES
#> 988  LEF1      SVEP1  0.54  STES
#> 989  LEF1       SVIL  0.40  STES
#> 990  LEF1     SWAP70  0.40  STES
#> 991  LEF1      SYDE1  0.59  STES
#> 992  LEF1      SYNE1  0.53  STES
#> 993  LEF1      SYNPO  0.47  STES
#> 994  LEF1     SYNPO2  0.47  STES
#> 995  LEF1      SYT11  0.61  STES
#> 996  LEF1      SYT15  0.41  STES
#> 997  LEF1      TACR1  0.46  STES
#> 998  LEF1      TAGAP  0.41  STES
#> 999  LEF1       TAL1  0.45  STES
#> 1000 LEF1    TBC1D2B  0.40  STES
#> 1001 LEF1      TBCEL  0.50  STES
#> 1002 LEF1     TCEAL7  0.55  STES
#> 1003 LEF1      TCF21  0.55  STES
#> 1004 LEF1       TCF4  0.65  STES
#> 1005 LEF1     TCF7L1  0.44  STES
#> 1006 LEF1    TCP11L1  0.40  STES
#> 1007 LEF1   TCTEX1D1  0.41  STES
#> 1008 LEF1        TEK  0.53  STES
#> 1009 LEF1       TFEC  0.41  STES
#> 1010 LEF1    TGFB1I1  0.50  STES
#> 1011 LEF1      TGFB2  0.44  STES
#> 1012 LEF1      TGFB3  0.49  STES
#> 1013 LEF1      TGFBI  0.42  STES
#> 1014 LEF1      THBS1  0.45  STES
#> 1015 LEF1      THBS2  0.46  STES
#> 1016 LEF1      THBS4  0.51  STES
#> 1017 LEF1      THSD1  0.42  STES
#> 1018 LEF1     THSD7A  0.47  STES
#> 1019 LEF1     THSD7B  0.52  STES
#> 1020 LEF1       THY1  0.56  STES
#> 1021 LEF1      TIMP2  0.51  STES
#> 1022 LEF1      TIMP3  0.54  STES
#> 1023 LEF1       TLL1  0.45  STES
#> 1024 LEF1       TLR4  0.49  STES
#> 1025 LEF1       TLR7  0.49  STES
#> 1026 LEF1     TM6SF1  0.54  STES
#> 1027 LEF1      TMCC2  0.52  STES
#> 1028 LEF1    TMEM108  0.47  STES
#> 1029 LEF1    TMEM119  0.61  STES
#> 1030 LEF1    TMEM130  0.50  STES
#> 1031 LEF1   TMEM132E  0.41  STES
#> 1032 LEF1   TMEM150C  0.51  STES
#> 1033 LEF1   TMEM200A  0.46  STES
#> 1034 LEF1   TMEM200B  0.44  STES
#> 1035 LEF1   TMEM200C  0.54  STES
#> 1036 LEF1    TMEM204  0.54  STES
#> 1037 LEF1     TMEM26  0.39  STES
#> 1038 LEF1     TMEM35  0.50  STES
#> 1039 LEF1     TMEM47  0.55  STES
#> 1040 LEF1     TMEM74  0.40  STES
#> 1041 LEF1      TMOD2  0.40  STES
#> 1042 LEF1    TMSB15A  0.44  STES
#> 1043 LEF1      TMTC1  0.39  STES
#> 1044 LEF1  TNFAIP8L3  0.49  STES
#> 1045 LEF1    TNFSF11  0.49  STES
#> 1046 LEF1     TNFSF4  0.46  STES
#> 1047 LEF1     TNFSF8  0.43  STES
#> 1048 LEF1        TNR  0.43  STES
#> 1049 LEF1     TNRC6C  0.41  STES
#> 1050 LEF1       TNS1  0.51  STES
#> 1051 LEF1       TOX2  0.42  STES
#> 1052 LEF1   TRAF3IP3  0.41  STES
#> 1053 LEF1       TRIL  0.45  STES
#> 1054 LEF1     TRIM61  0.44  STES
#> 1055 LEF1        TRO  0.57  STES
#> 1056 LEF1      TRPC1  0.54  STES
#> 1057 LEF1      TSHZ2  0.54  STES
#> 1058 LEF1      TSHZ3  0.59  STES
#> 1059 LEF1    TSPAN11  0.57  STES
#> 1060 LEF1    TSPAN18  0.45  STES
#> 1061 LEF1     TSPAN2  0.50  STES
#> 1062 LEF1     TSPAN7  0.41  STES
#> 1063 LEF1     TSPAN9  0.42  STES
#> 1064 LEF1     TSPYL2  0.38  STES
#> 1065 LEF1     TSPYL5  0.44  STES
#> 1066 LEF1      TTC28  0.53  STES
#> 1067 LEF1        TUB  0.47  STES
#> 1068 LEF1     TUBA1A  0.57  STES
#> 1069 LEF1      TUSC3  0.47  STES
#> 1070 LEF1     TWIST2  0.53  STES
#> 1071 LEF1      TXLNB  0.56  STES
#> 1072 LEF1     UBE2E2  0.39  STES
#> 1073 LEF1      UNC5C  0.45  STES
#> 1074 LEF1      USP44  0.42  STES
#> 1075 LEF1      USP51  0.52  STES
#> 1076 LEF1       UTRN  0.42  STES
#> 1077 LEF1      VASH1  0.51  STES
#> 1078 LEF1      VASH2  0.48  STES
#> 1079 LEF1      VCAM1  0.51  STES
#> 1080 LEF1       VCAN  0.52  STES
#> 1081 LEF1      VEGFC  0.56  STES
#> 1082 LEF1      VGLL3  0.49  STES
#> 1083 LEF1        VIM  0.56  STES
#> 1084 LEF1      VIPR2  0.47  STES
#> 1085 LEF1        VWF  0.38  STES
#> 1086 LEF1        WAS  0.39  STES
#> 1087 LEF1      WASF3  0.39  STES
#> 1088 LEF1    WBSCR17  0.48  STES
#> 1089 LEF1      WDFY4  0.45  STES
#> 1090 LEF1      WIPF1  0.57  STES
#> 1091 LEF1      WISP1  0.46  STES
#> 1092 LEF1      WNT2B  0.52  STES
#> 1093 LEF1      WNT9B  0.42  STES
#> 1094 LEF1      WSCD1  0.45  STES
#> 1095 LEF1       WWC3  0.44  STES
#> 1096 LEF1     ZBTB20  0.46  STES
#> 1097 LEF1    ZCCHC18  0.56  STES
#> 1098 LEF1    ZCCHC24  0.59  STES
#> 1099 LEF1    ZDHHC15  0.57  STES
#> 1100 LEF1       ZEB1  0.59  STES
#> 1101 LEF1       ZEB2  0.60  STES
#> 1102 LEF1      ZFHX4  0.45  STES
#> 1103 LEF1      ZFP28  0.51  STES
#> 1104 LEF1      ZFP82  0.48  STES
#> 1105 LEF1      ZFPM2  0.58  STES
#> 1106 LEF1      ZMAT1  0.40  STES
#> 1107 LEF1     ZNF135  0.63  STES
#> 1108 LEF1      ZNF25  0.48  STES
#> 1109 LEF1     ZNF333  0.45  STES
#> 1110 LEF1    ZNF354C  0.48  STES
#> 1111 LEF1     ZNF366  0.43  STES
#> 1112 LEF1    ZNF385B  0.42  STES
#> 1113 LEF1    ZNF385D  0.42  STES
#> 1114 LEF1     ZNF423  0.66  STES
#> 1115 LEF1      ZNF43  0.41  STES
#> 1116 LEF1     ZNF439  0.42  STES
#> 1117 LEF1     ZNF454  0.56  STES
#> 1118 LEF1     ZNF470  0.42  STES
#> 1119 LEF1     ZNF471  0.51  STES
#> 1120 LEF1     ZNF521  0.61  STES
#> 1121 LEF1     ZNF532  0.40  STES
#> 1122 LEF1     ZNF568  0.51  STES
#> 1123 LEF1     ZNF569  0.44  STES
#> 1124 LEF1     ZNF570  0.47  STES
#> 1125 LEF1     ZNF582  0.53  STES
#> 1126 LEF1     ZNF583  0.41  STES
#> 1127 LEF1     ZNF610  0.49  STES
#> 1128 LEF1     ZNF625  0.43  STES
#> 1129 LEF1     ZNF626  0.49  STES
#> 1130 LEF1     ZNF660  0.52  STES
#> 1131 LEF1     ZNF677  0.56  STES
#> 1132 LEF1      ZNF71  0.40  STES
#> 1133 LEF1     ZNF781  0.53  STES
#> 1134 LEF1    ZNF804A  0.45  STES
#> 1135 LEF1     ZNF829  0.46  STES
#> 1136 LEF1     ZNF835  0.50  STES
#> 1137 LEF1     ZNF843  0.47  STES
#> 1138 LEF1     ZNF853  0.53  STES
#> 1139 LEF1     ZNF880  0.42  STES
#> 1140 LEF1    ZSCAN18  0.46  STES

# enter a custom query with different arguments
get_tf(conn,
       tf = 'LEF1',
       study = 'STES',
       min_abs_cor = .3,
       max_num = 5)
#>     tf  feature  cor study
#> 1 LEF1    ENOX1 0.68  STES
#> 2 LEF1 KIAA1755 0.67  STES
#> 3 LEF1   OLFML1 0.66  STES
#> 4 LEF1    RFTN1 0.66  STES
#> 5 LEF1   ZNF423 0.66  STES